% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_clusters_wgcna.R
\name{add_clusters_wgcna}
\alias{add_clusters_wgcna}
\title{Function to add the clusters obtained from wgcna}
\usage{
add_clusters_wgcna(object, which_data, baseline, ...)
}
\arguments{
\item{object}{An S4 object of class metime_analyser}

\item{which_data}{character to define which dataset is to be used}

\item{baseline}{character to define the timepoint to be used as baseline to predict clusters}

\item{...}{other parameters for cutreeDynamic function such minClusterSize, pamDendroRespect etc}
}
\value{
metime_analyser object with updated column info about the clustersize
}
\description{
A function to add cluster assiginment to the col_data from WGCNA
}
