% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_add.R
\name{add_col_stats}
\alias{add_col_stats}
\title{Function to check normality and add data to col data}
\usage{
add_col_stats(object, which_data, type, metab_names, all)
}
\arguments{
\item{object}{An object of class metime_analyser}

\item{which_data}{dataset on which the method is to be applied}

\item{type}{type of test, "shapiro" and "kruskal" are available}

\item{metab_names}{column that has the metabolite names in col_data.}

\item{all}{logical to add all kinds of available stats.}
}
\value{
S4 object with shapiro wilk test related data in the col_data
}
\description{
A method applied on the s4 object of class "metime_analyser" to check normality of the metabolites
and add it to corresponding columns
}
\examples{
object <- add_col_normality(object=data, which_data=c("lipid_data","nmr_data"), type="shapiro", metab_names=c("metabolite","id"))
}
