% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_add.R
\name{add_covariates_for_ggms}
\alias{add_covariates_for_ggms}
\title{Function to add covariates to the dataset of interest for GGMs}
\usage{
add_covariates_for_ggms(object, which_data, covariates, class.ind, phenotype)
}
\arguments{
\item{object}{object of class metime_analyser}

\item{which_data}{Dataset/s to which the covariates is to be added}

\item{covariates}{list of character vector names of covariates.
Use based on number of dataset cov <- list(c(covs_lipids), c(covs_nmr))}

\item{class.ind}{Logical to convert factor variables into class.ind style or not}

\item{phenotype}{Logical. If True will extract from phenotype dataset else uses row data}
}
\value{
S4 object with covariates added to the dataset
}
\description{
adds Covariates to data matrices in metime_analyser S4 object
}
