% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dataset.R
\name{add_dataset}
\alias{add_dataset}
\title{This function appends an object of class metime_analyser with a new dataset.}
\usage{
add_dataset(object, data, col_data, row_data, name)
}
\arguments{
\item{object}{S4 object of class metime_analyser}

\item{data}{data.frame containing data}

\item{col_data}{data.frame containing col_data: id column of col data has to match colnames of data}

\item{row_data}{data.frame containing row_data: id column of row data has to match rownames of data}

\item{name}{Name of the new dataset}
}
\value{
An object of class metime_analyser
}
\description{
function to apply on metime_analyser object to append a new dataset into the existing object
}
\examples{
# append data frames into the metime_analyser object
appended_object <- add_dataset(object=metime_analyser_object, 
                   data=data, row_data=row_data, col_data=col_data, name="name of the new dataset")
}
