% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_add.R
\name{add_distribution_vars_to_rows}
\alias{add_distribution_vars_to_rows}
\title{Function to add measurements taken at screening time for samples to be added to all timepoints in row data}
\usage{
add_distribution_vars_to_rows(
  object,
  screening_vars,
  distribution_vars,
  which_data
)
}
\arguments{
\item{object}{An object of class metime_analyser}

\item{which_data}{dataset to which the information is to be added(only 1 can be used at a time)}

\item{vars}{A character naming the vars of interest}
}
\value{
object of class metime_analyser with phenotype data added to row data
}
\description{
A method applied on the s4 object of class "metime_analyser" to add all those datapoints that were measured only during screening
to all the respective samples at all timepoints in row_data lists
}
\examples{
# adding APOEGrp, PTGENDER, and diag group to all data points and prepping the object for viz_distribution_plotter()
object <- add_distribution_vars_to_rows(object=data, screening_vars=c("APOEGrp", "DXGrp_longi", "PTGENDER"), 
		distribution_vars=c("Age", "BMI", "ADNI_MEM", "ADNI_LAN", "ADNI_EF", "APOEGrp", "DXGrp_longi", "PTGENDER"), which_data="lipid_data")
}
