% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_function_info.R
\name{add_function_info}
\alias{add_function_info}
\title{Function to add information of function added to the data}
\usage{
add_function_info(object, function_name, params)
}
\arguments{
\item{object}{S4 object of class metime_analyser}

\item{function_name}{name of the function used}

\item{params}{other parameters used wrapped as a list}
}
\value{
object of class metime_analyser with the information of method applied
}
\description{
Function to add information about the method applied to the dataset
}
