% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_metabs_as_covariates.R
\name{add_metabs_as_covariates}
\alias{add_metabs_as_covariates}
\title{Add columns from data x to data y within a metime_analyzer.}
\usage{
add_metabs_as_covariates(object, which_data, which_metabs)
}
\arguments{
\item{object}{a S4 object of the class "metime_analyzer".}

\item{which_data}{a character defining the dataset to which the metab data is to be added.}

\item{which_metabs}{a named list of character vectors. Names of the list elements correspond to the name of dataset, the character vectors define columns to be used. eg: which_metabs=list(nmr_data=c("metab1", "metab2"), lipid_data=c("metab3", "metab4"))}
}
\value{
object of class metime_analyser with columns from one dataset added for analysis to another dataset
}
\description{
Merge a specified columns from datasets within the S4 object to one data.
}
