% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_add.R
\name{add_node_features}
\alias{add_node_features}
\title{Function to add features to visnetwork plot from another plotter object}
\usage{
add_node_features(
  network_plotter_object,
  guide_plotter_object,
  which_type,
  metab_colname
)
}
\arguments{
\item{network_plotter_object}{plotter object with network information}

\item{guide_plotter_object}{guide from which the colors are to be extracted}

\item{which_type}{type of the guide plotter object to be used. Current options are "regression" and "conservation"}

\item{metab_colname}{name of the column in guide plotter object that represents the metabolites}
}
\value{
network plotter object with new node colors/features
}
\description{
Function to add node features to see the nodes in the network that affected differently
}
