% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_phenotypes_as_covariates.R
\name{add_phenotypes_as_covariates}
\alias{add_phenotypes_as_covariates}
\title{Function to add covariates to the dataset of interest for GGMs}
\usage{
add_phenotypes_as_covariates(
  object,
  which_data,
  covariates,
  class.ind,
  phenotype
)
}
\arguments{
\item{object}{object of class metime_analyser}

\item{which_data}{Dataset to which the covariates is to be added}

\item{covariates}{character vector names of covariates.}

\item{class.ind}{Logical to convert factor variables into class.ind style or not}

\item{phenotype}{Logical. If True will extract from phenotype dataset else uses row data}
}
\value{
S4 object with covariates added to the dataset
}
\description{
adds Covariates to data matrices in metime_analyser S4 object
}
