% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plot.R
\name{add_plot}
\alias{add_plot}
\title{Generates a report from an metime_analyser object}
\usage{
add_plot(
  object,
  result_item = NULL,
  plot_data = NULL,
  calc_type = NA,
  calc_info = NA,
  result_index = "result_index",
  result_type = "result_type",
  result_subtype = "result_subtype",
  result_title = "result_title"
)
}
\arguments{
\item{object}{a S4 object of class metime_analyser}

\item{result_item}{a plot or data table to be added as a result.}

\item{plot_data}{a dataframe corresponding to the result item.}

\item{calc_type}{a character defining the calc type of an object that has been added. Default set to NA.}

\item{calc_info}{a character defining the calc info of an object that has been added. Default set to NA.}

\item{result_index}{a character defining the result index of an object that has been added. Default set to 'result_index'}

\item{result_type}{a character defining the result type of an object that has been added. Default set to 'result_type'}

\item{result_subtype}{a character defining the result subtype of an object that has been added. Default set to 'result_subtype'}

\item{result_title}{a character defining the result title of an object that has been added. Default set to 'result_title'}
}
\value{
a S4 object of class metime_analyser
}
\description{
Write an HTML or PDF report that summarizes one or more "me_time_analyser" objects and display all results
}
