% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_result.R
\name{add_result}
\alias{add_result}
\title{Add a result element to a metime_analyser}
\usage{
add_result(
  object,
  x,
  name = NULL,
  functions_applied = NULL,
  calc_info = NULL,
  calc_type = NULL,
  results_index = NULL
)
}
\arguments{
\item{object}{a S4 object of class "metime_analyser"}

\item{x}{a data.frame or plot that should be added to the list of results}

\item{functions_applied}{a character providing information on the calculation}

\item{calc_info}{a character that provides information on the calculation}

\item{calc_type}{a character that defines the type of calculation}

\item{results_index}{a haracter or numeric of length=1 to define the results to which you want to add x.
Default is set to NULL and that implies that new results are generated}
}
\value{
S4 object of class "metime_analyser" with appended results
}
\description{
Add a result from a calc_ function to the metime_analyser.
}
