% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_screening_vars.R
\name{add_screening_vars}
\alias{add_screening_vars}
\title{Add values measured at a screening time for samples to be added to all time points}
\usage{
add_screening_vars(object, vars)
}
\arguments{
\item{object}{a S4 object of class "metime_analyser".}

\item{vars}{a character vector naming the columns of interest}
}
\value{
S4 object of class "metime_analyser" with screening measurements added to all other time points of the same subject.
}
\description{
Add all data points that were measured only during a defined screening time point to all other measurements of the same subject.
Note that this function is only valid in the case of phenotype data and not in the case of row_data
}
