% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_general.R
\name{adni_add_index}
\alias{adni_add_index}
\title{Function to add indices i.e. RID and timepoints to data matrix}
\usage{
adni_add_index(data)
}
\arguments{
\item{data}{data matrix with rownames as adni_ids}
}
\value{
data matrix with new added columns
}
\description{
add ids to check for data with ease
}
