% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_general.R
\name{adni_filter_full_tp}
\alias{adni_filter_full_tp}
\title{Function to extract timepoints of ineterest for longitudnal analysis}
\usage{
adni_filter_full_tp(x.data, tp, by = c("timepoint"))
}
\arguments{
\item{x.data}{data matrix of interest}

\item{tp}{timepoints of interest}
}
\value{
data matrix with only timepoints of interest
}
\description{
function to extract the timepoints that are of importance in longitudnal analysis similar to split_acc_to_time except is applied on data matrix rather than the object itself
}
