% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_ggms.R
\name{adni_ggm_calc_ggm_dynamic}
\alias{adni_ggm_calc_ggm_dynamic}
\title{Function to calculate a dynamic GeneNet GGM}
\usage{
adni_ggm_calc_ggm_dynamic(data, threshhold = c("bonferoni", "FDR", "li"))
}
\arguments{
\item{data}{data matrix in a longitudnal format(see adni_convert_ggm_longitudnal)}

\item{threshold}{type of multiple hypothesis correction. Available are Bonferoni("bonferoni"), Benjamini-Hochberg("FDR") and independent tests method("li", also see Li et al ....)}
}
\value{
a dataframe with edges, partial correlation and associated p-values
}
\description{
calculates GGM on longitudnal data matrix and returns a dataframe with edges, partial correlation and associated p-values
}
