% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_ggms.R
\name{adni_ggm_convert_longitudinal}
\alias{adni_ggm_convert_longitudinal}
\title{Function to convert data into a longitudnal format for GeneNet ggms}
\usage{
adni_ggm_convert_longitudinal(data)
}
\arguments{
\item{data}{data matrix with metabolite concentrations}
}
\value{
data matrix converted into a longitudnal format
}
\description{
converts a dataset with full data into a longitudnal version
}
