% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_general.R
\name{adni_independent_tests}
\alias{adni_independent_tests}
\title{Function to filter out significant edges from genenet ggm}
\usage{
adni_independent_tests(data)
}
\arguments{
\item{data}{data matrix with rownames as adni_ids}
}
\value{
data frame with manipulated eigenvalues that can be used to filter significant edges
}
\description{
Function to filter out significant edges from genenet ggm
}
