% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_general.R
\name{adni_rm_index}
\alias{adni_rm_index}
\title{Function to remove indices i.e. RID and timepoints from data matrix}
\usage{
adni_rm_index(data)
}
\arguments{
\item{data}{data matrix with rownames as adni_ids}
}
\value{
data matrix without index columns
}
\description{
remove ids that are used to check for data with ease(see adni_add_index())
}
