% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_ggms.R
\name{automated_ggm_genenet}
\alias{automated_ggm_genenet}
\title{An automated fucntion to calculate GGM from genenet}
\usage{
automated_ggm_genenet(object, which_data, threshhold, timepoints)
}
\arguments{
\item{object}{object obtained after applying prep_data_for_ggms() on S4 object of class metab_analyse with type="single"}

\item{which_data}{a character or a character vector naming the datasets of interest}

\item{timepoints}{timepoints of interest that are to be used to build networks}

\item{threshold}{type of threshold to be used for extracting significant edges}
}
\value{
Network data with edgelist, partial correlation values and associated p-values and corrected p-values
}
\description{
automated funtion that can be applied on s3 object obtained after prep_data_for_ggms() to obtain geneNet network along with threshold used
}
