% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_ggms.R
\name{automated_ggm_mlp}
\alias{automated_ggm_mlp}
\title{An automated fucntion to calculate GGM from genenet}
\usage{
automated_ggm_mlp(object, which_data, rho, nfolds, timepoints)
}
\arguments{
\item{object}{object obtained after applying prep_data_for_ggms() on S4 object of class metab_analyse with type="multi" not applicable to apply on singular sets}

\item{which_data}{a character or a character vector naming the datasets of interest}

\item{rho}{tuning parameter for regression}

\item{nfolds}{check}

\item{timepoints}{timepoints of interest that are to be used to build networks}
}
\value{
Network data with edgelist, partial correlation values and associated p-values and corrected p-values
}
\description{
automated funtion that can be applied on s3 object obtained after prep_data_for_ggms() to obtain geneNet network along with threshold used.
This function is not applicable for singular datasets
}
