% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_ggms.R
\name{automated_temporal_network}
\alias{automated_temporal_network}
\title{An automated function to caluclate temporal network with lagged model}
\usage{
automated_temporal_network(object, lag, timepoints)
}
\arguments{
\item{object}{object obtained after applying prep_data_for_ggms() on S4 object of class metab_analyser}

\item{lag}{which lagged model to use. 1 means one-lagged model, similary 2,3,..etc}

\item{timepoints}{timepoints of interest that are to be used to build networks}
}
\value{
temporal network data with edgelist and regression values
}
\description{
calculates temporal networks for each dataset with a lagged model as used in graphical VAR
}
