% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_calc.R
\name{calc_boruta}
\alias{calc_boruta}
\title{Function to calculate dependent variables}
\usage{
calc_boruta(object, which_x, which_y, verbose, output_loc, file_name)
}
\arguments{
\item{object}{An object of class metime_analyser}

\item{which_x}{Name of the dataset to be used for training}

\item{which_y}{Name of the dataset to be used for testing}

\item{verbose}{Information provided on steps being processed}

\item{output_loc}{path to the parent directory where in the out file wíll be stored}

\item{file_name}{name of the out file}
}
\value{
List of conservation index results
}
\description{
An S4 method to be applied on the metime_analyser object so as to calculate dependent variables
}
