% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_clusters_wgcna.R
\name{calc_clusters_wgcna}
\alias{calc_clusters_wgcna}
\title{Calculate cluster assignment from WGCNA}
\usage{
calc_clusters_wgcna(
  object,
  which_data,
  baseline = "t0",
  cols_for_meta = NULL,
  name = "WGCNA_clusters_1",
  soft_power = NULL,
  ...
)
}
\arguments{
\item{object}{a S4 object of the class metime_analyzer.}

\item{which_data}{a character to define which dataset is to be used.}

\item{baseline}{a character to define the baseline time point which is used for cluster calculation.}

\item{cols_for_meta}{A list of named character vector of length equal to which_data.
Example: list(dataset1=c(new_name1="colname", new_name2="colname"))
Default is set to NULL}

\item{name}{Name of the results. Default is set to "WGCNA_clusters_1"}

\item{soft_power}{Numeric soft-thresholding power to use. When \code{NULL}, the
first candidate power with a truncated R-squared greater than 0.8 is used.
If no candidate meets that threshold, the function warns and returns without
changing \code{object}. Supply a numeric value to override the automatic choice.}

\item{...}{multiple parameters separated by commas passed to cutreeDynamic (R package: dynamicTreeCut).
Parameters include: minClusterSize, pamDendroRespect ...}
}
\value{
a S4 object of class "metime_analyser" with cluster information appended to col_data of which_data
}
\description{
calculate cluster assignment to specified col_data from Weighted Correlation Network Analysis
}
\details{
Based on the method described in the WGCNA tutorials for step-by-step network construction and module detection. The idea is to find
modules of metabolites at baseline. This method can be found in detail here \href{https://bmcbioinformatics.biomedcentral.com/articles/10.1186/1471-2105-9-559}{WGCNA: an R package for weighted correlation network analysis}
}
\seealso{
\link[dynamicTreeCut:cutreeDynamic]{dynamicTreeCut::cutreeDynamic}, \link{get_metadata_for_columns}, \link{mod_trans_eigendata}
}
