% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_col_stats.R
\name{calc_col_stats}
\alias{calc_col_stats}
\title{Check normality of metabolites and append results to col_data}
\usage{
calc_col_stats(
  object,
  which_data,
  method = c("shapiro", "ks"),
  cols_for_meta,
  name = "calc_col_stats_1"
)
}
\arguments{
\item{object}{a S4 object of the class "metime_analyzer".}

\item{which_data}{a character to define which dataset is to be used.}

\item{method}{a character vector specifying the method to check for normality. By default all available options are chosen: c("shapiro", "ks")". See details for more information.}

\item{cols_for_meta}{a list of named character vectors for obtaining metainformation of metabolites}

\item{name}{name of the results. Default is set to "calc_col_stats_1"}
}
\value{
a  S4 object of class "metime_analyser" with results appended to the col_data of which_data
}
\description{
A method applied to data within the S4 object of class "metime_analyser" to check normality of the metabolites
and generate a results section for easy visualization,
}
\details{
Shapiro: Performs the Shapiro-Wilk test of normality (normal distribution if shapiro_pval > 0.05). This test is based on correlation between the data and the corresponding normal scores.
Ks: Performs a one-sample Kolmogorov-Smirnov test (normal distribution if ks_pval > 0.05). The ks.test compares the metabolite values against pnorm.
}
\seealso{
\link{shapiro.test} \link{ks.test} \link{get_metadata_for_columns}
}
