% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_colinearity.R
\name{calc_colinearity}
\alias{calc_colinearity}
\title{Function to calculate colinearity}
\usage{
calc_colinearity(
  object,
  which_data,
  cols_for_meta,
  show_all = F,
  name = "calc_colinearity_1",
  stratifications = list()
)
}
\arguments{
\item{object}{An S4 object of class metime_analyser}

\item{which_data}{Dataset to check for colinearity}

\item{cols_for_meta}{list of character vectors of column names needed in metadata. Id and class name is needed}

\item{show_all}{logical. True will only filter out colinear data}

\item{name}{character to define the name of the result}

\item{stratifications}{List to stratify data into a subset. Usage list(name=values)}
}
\value{
plotter object with data for heatmap information
}
\description{
Function to calculate colinearity in a dataset
}
