% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_colinearity_features.R
\name{calc_colinearity_features}
\alias{calc_colinearity_features}
\title{Function to calculate colinearity of features}
\usage{
calc_colinearity_features(
  object,
  which_data,
  name = "calc_colinearity_1",
  stratifications = list()
)
}
\arguments{
\item{object}{An S4 object of class metime_analyser}

\item{which_data}{Dataset to check for colinearity}

\item{name}{character to define the name of the result}

\item{stratifications}{List to stratify data into a subset. Usage list(name=values)}

\item{show_all}{logical. True will only filter out colinear data}
}
\value{
plotter object with data for heatmap information
}
\description{
Function to calculate colinearity in a dataset
}
