% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_conservation_metabotype.R
\name{calc_conservation_metabotype}
\alias{calc_conservation_metabotype}
\title{Function to calculate metabotype conservation index}
\usage{
calc_conservation_metabotype(
  object,
  which_data,
  verbose = F,
  cols_for_meta,
  stratifications,
  name = "calc_conservation_metabotype_1"
)
}
\arguments{
\item{object}{An object of class metime_analyser}

\item{which_data}{Name of the dataset to be used}

\item{verbose}{Information provided on steps being processed}

\item{cols_for_meta}{Character vector to define column names that are to be used for plotting purposes}

\item{stratifications}{List to stratify data into a subset. Usage list(name=value)}

\item{name}{character vector to define the results}
}
\value{
List of conservation index results
}
\description{
Method applied on the object metime_analyser to calculate the metabotype conservation index
}
\examples{
#calculating metabotype_conservation_index 
out <- calc_metabotype_conservation(object=metime_analyser_object, which_data="Name of the dataset")
}
