% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_calc.R
\name{calc_correlation}
\alias{calc_correlation}
\title{Function to calculate correlation}
\usage{
calc_correlation(object, which_data, method)
}
\arguments{
\item{object}{S4 Object of class metime_analyser}

\item{which_data}{specify datasets to calculate on. One or more possible}

\item{method}{default setting: method="pearson", Alternative "spearman" also possible}
}
\value{
data.frame with pairwise results
}
\description{
calculate pairwise correlations
This function creates a dataframe for plotting from a dataset.
}
\examples{
# Example to calculate correlations
dist <- calc_correlation(object=metime_analyser_object, which_data="name of the dataset", 
          method="pearson")
}
