% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_correlation_features.R
\name{calc_correlation_features}
\alias{calc_correlation_features}
\title{Function to calculate correlation}
\usage{
calc_correlation_features(
  object,
  which_data,
  method,
  name = "calc_correlation_pairwise_1",
  stratifications
)
}
\arguments{
\item{object}{S4 Object of class metime_analyser}

\item{which_data}{specify datasets to calculate on. One or more possible}

\item{method}{default setting: method="pearson", Alternative "spearman" also possible}

\item{name}{name of the results should be of length=1}

\item{stratifications}{List to stratify data into a subset. Usage list(name=value)}
}
\value{
data.frame with pairwise results
}
\description{
calculate pairwise correlations
This function creates a dataframe for plotting from a dataset.
}
\examples{
# Example to calculate correlations
dist <- calc_correlation(object=metime_analyser_object, which_data="name of the dataset", 
          method="pearson")
}
