% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dimensionality_reduction.R
\name{calc_dimensionality_reduction}
\alias{calc_dimensionality_reduction}
\title{Function to calculate dimensionality reduction methods such as tsne, umap and pca.}
\usage{
calc_dimensionality_reduction(
  object,
  which_data,
  type,
  cols_for_metabs,
  cols_for_samples,
  stratifications,
  ...
)
}
\arguments{
\item{object}{An object of class metime_analyser}

\item{which_data}{a character vector - Names of the dataset from which the samples will be extracted}

\item{type}{type of the dimensionality reduction method to be applied. Accepted inputs are "UMAP", "tSNE", "PCA"}

\item{cols_for_metabs}{a list of character vectors for getting metadata for columns for plotting purposes}

\item{cols_for_samples}{a character vector to define the columns to extract metadata for plotting purposes}

\item{stratifications}{List to stratify data into a subset. Usage list(name=value)}

\item{...}{additional arguments that can be passed on to prcomp(), M3C::tsne() and umap::umap()}
}
\value{
a list with two plotter objects containing the dimensionality reduction information that can be parsed into plotting function
1) samples - data of the individuals(".$samples")
2) metabs - data of the metabolites(".$metabs")
}
\description{
A method to apply on s4 object of class metime_analyse in order to obtain information after dimensionality reduction on a dataset/s
}
\examples{
#calculate PCA
pca <- calc_dimensionality_reduction(object=metime_analyser_object, which_data="name/s of the dataset/s", type="PCA")
#calculate UMAP
pca <- calc_dimensionality_reduction(object=metime_analyser_object, which_data="name/s of the dataset/s", type="UMAP")
#calculate tSNE
pca <- calc_dimensionality_reduction(object=metime_analyser_object, which_data="name/s of the dataset/s", type="tSNE")
}
