% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_distance_samples.R
\name{calc_distance_samples}
\alias{calc_distance_samples}
\title{Function to calculate dissimilarity using distance measures}
\usage{
calc_distance_samples(
  object,
  which_data,
  method,
  name = "calc_distance_pairwise_1",
  stratifications
)
}
\arguments{
\item{object}{S4 Object of class metime_analyser}

\item{which_data}{specify datasets to calculate on. One or more possible}

\item{method}{default setting: method="euclidean", Alternative "maximum","minimum",
"manhattan","canberra","minkowski" are also possible}

\item{name}{name of the results should be of length=1}

\item{stratifications}{List to stratify data into a subset. Usage list(name=value)}
}
\value{
data.frame with pairwise results
}
\description{
calculate pairwise distances between samples
This function creates a dataframe for plotting from a dataset.
}
\examples{
# Example to calculate pairwise distances
dist <- calc_pairwise_distance(object=metime_analyser_object, which_data="name of the dataset", 
          method="euclidean")
}
