% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_distribution_*.R
\name{calc_distribution_metabs}
\alias{calc_distribution_metabs}
\title{Function for Plotting distributions of phenotypic variables}
\usage{
calc_distribution_metabs(
  object,
  which_data,
  cols,
  name = "calc_distribution_metabs_1"
)
}
\arguments{
\item{object}{An object of class metime_analyser}

\item{which_data}{A Character vector to define the dataset of interest}

\item{name}{Character to name the results}

\item{col_metabs}{character vector to define the columns whose distributions are wanted from col_data#' @param which_data Name of the dataset from which the samples will be extracted}
}
\value{
S4 object with updated plot_data and plots
with plots being either 1) density plot
2) bar plot and a line plot
}
\description{
A method to be applied onto s4 object so as to obtain distributions of various phenotypic variables
}
