% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_distribution_*.R
\name{calc_distribution_samples}
\alias{calc_distribution_samples}
\title{Function for Plotting distributions of phenotypic variables}
\usage{
calc_distribution_samples(
  object,
  which_data,
  cols,
  stratifications,
  name = "calc_distribution_samples_1"
)
}
\arguments{
\item{object}{An object of class metime_analyser}

\item{which_data}{Name of the dataset from which the samples will be extracted}

\item{cols}{character vector to define the columns whose distributions are wanted from row_data}

\item{stratifications}{List to define the stratification of interest}

\item{name}{Character to name the results}
}
\value{
S4 object with updated plot_data and plots
with plots being either 1) density plot
2) bar plot and a line plot
}
\description{
A method to be applied onto s4 object so as to obtain distributions of various phenotypic variables
}
