% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_calc.R
\name{calc_ggm_genenet}
\alias{calc_ggm_genenet}
\title{An automated fucntion to calculate GGM from genenet}
\usage{
calc_ggm_genenet(object, which_data, threshold, timepoints)
}
\arguments{
\item{object}{S4 object of cĺass metime_analyser}

\item{which_data}{a character or a character vector naming the datasets of interest}

\item{threshold}{type of threshold to be used for extracting significant edges}

\item{timepoints}{timepoints of interest that are to be used to build networks(as per timepoints in rows)}
}
\value{
Network data with edgelist, partial correlation values and associated p-values and corrected p-values
}
\description{
automated funtion that can be applied on s3 object obtained after prep_data_for_ggms() to obtain geneNet network along with threshold used
}
