% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ggm_genenet_crosssectional.R
\name{calc_ggm_genenet_crosssectional}
\alias{calc_ggm_genenet_crosssectional}
\title{An automated fucntion to calculate GGM from genenet crosssectional version}
\usage{
calc_ggm_genenet_crosssectional(
  object,
  which_data,
  threshold,
  timepoint,
  all,
  cols_for_meta,
  covariates,
  ...
)
}
\arguments{
\item{object}{S4 object of cĺass metime_analyser}

\item{which_data}{a character or a character vector naming the datasets of interest}

\item{threshold}{type of threshold to be used for extracting significant edges.
allowed inputs are "li", "FDR", "bonferroni"}

\item{all}{Logical to extract all edges without any pval correction}

\item{cols_for_meta}{list of character vector for extracting metadata of metabolites for plotting}

\item{covariates}{covariates to be used for this analysis}

\item{...}{additional arguments for GeneNet}

\item{timepoints}{timepoints of interest that are to be used to build networks(as per timepoints in rows)}
}
\value{
Network data as a plotter object
}
\description{
automated funtion that can be applied on metime_analyser object to obtain geneNet network along with threshold used
}
