% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_calc.R
\name{calc_ggm_genenet_longitudnal}
\alias{calc_ggm_genenet_longitudnal}
\title{An automated fucntion to calculate GGM from genenet longitudnal version}
\usage{
calc_ggm_genenet_longitudnal(
  object,
  which_data,
  threshold,
  timepoints,
  all,
  cols_for_meta,
  covariates,
  ...
)
}
\arguments{
\item{object}{S4 object of cĺass metime_analyser}

\item{which_data}{a character or a character vector naming the datasets of interest}

\item{threshold}{type of threshold to be used for extracting significant edges}

\item{timepoints}{timepoints of interest that are to be used to build networks(as per timepoints in rows)}

\item{all}{Logical to get all edges without any cutoff.}

\item{cols_for_meta}{a list of character vectors for getting metadata for columns for plotting purposes}

\item{covariates}{covariates to be used for this analysis}

\item{...}{addtional arguments for genenet network}
}
\value{
Network data as a plotter object
}
\description{
automated funtion that can be applied on metime_analyser object to obtain geneNet network along with threshold used
}
