% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_calc.R
\name{calc_li_thresh}
\alias{calc_li_thresh}
\title{Function to calculate multiple tests using method described by li}
\usage{
calc_li_thresh(object, which_data)
}
\arguments{
\item{object}{an S4 object of class metime_analyser}

\item{which_data}{dataset to be used for testing}
}
\value{
li threshold value
}
\description{
li test to check for colinearlity and use it for feature selection
}
