% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_lm_MatrixEQTL.R
\name{calc_lm_matrixeqtl}
\alias{calc_lm_matrixeqtl}
\title{Function to perform matrixEQTL style regression longitudinally}
\usage{
calc_lm_matrixeqtl(
  object,
  which_data,
  covariates,
  feature_selection,
  regression_phenotype,
  name,
  stratifications,
  cols_for_meta
)
}
\arguments{
\item{object}{An S4 object of class metime_analyser}

\item{which_data}{character to define dataset to be used}

\item{covariates}{covariates to be used}

\item{feature_selection}{results from feature_selection}

\item{regression_phenotype}{character to define which phenotype}

\item{name}{character vector for naming the results}

\item{stratifications}{List to stratify data into a subset. Usage list(name=value)}

\item{cols_for_meta}{colnames to be added for meta information. list of character vectors of length which_data}
}
\value{
plotter object with a forest plot
}
\description{
Function to perform matrixEQTL style regression longitudinally
}
