% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_calc.R
\name{calc_metabolite_conservation}
\alias{calc_metabolite_conservation}
\title{Function to calculate metabolite conservation index}
\usage{
calc_metabolite_conservation(object, which_data, verbose)
}
\arguments{
\item{object}{An object of class metab_analyser}

\item{which_data}{Name of the dataset to be used}

\item{verbose}{Information provided on steps being processed}
}
\value{
List of conservation index results
}
\description{
Function to calculate metabolite conservation index
}
