% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_calc.R
\name{calc_metabotype_conservation}
\alias{calc_metabotype_conservation}
\title{Function to calculate metabotype conservation index}
\usage{
calc_metabotype_conservation(object, which_data, verbose)
}
\arguments{
\item{object}{An object of class metime_analyser}

\item{which_data}{Name of the dataset to be used}

\item{verbose}{Information provided on steps being processed}
}
\value{
List of conservation index results
}
\description{
Method applied on the object metime_analyser to calculate the metabotype conservation index
}
\examples{
#calculating metabotype_conservation_index 
out <- calc_metabotype_conservation(object=metime_analyser_object, which_data="Name of the dataset")
}
