% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_calc.R
\name{calc_pairwise_distance}
\alias{calc_pairwise_distance}
\title{Function to calculate dissimilarity using distance measures}
\usage{
calc_pairwise_distance(object, which_data, method)
}
\arguments{
\item{object}{S4 Object with}

\item{which_data}{specify datasets to calculate on. One or more possible}

\item{method}{default setting: method="euclidean", Alternative "maximum","minimum",
"manhattan","canberra","minkowski" are also possible}
}
\value{
data.frame with pairwise results
}
\description{
calculate pairwise distances
This function creates a dataframe for plotting from a dataset.
}
\examples{
# Example to calculate pairwise distances
dist <- calc_pairwise_distance(object=metab_analyser_object, which_data="name of the dataset", 
          method="euclidean")
}
