% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_temporal_ggm.R
\name{calc_temporal_ggm}
\alias{calc_temporal_ggm}
\title{An automated function to caluclate temporal network with lagged model}
\usage{
calc_temporal_ggm(
  object,
  which_data,
  lag,
  timepoints,
  alpha,
  nfolds,
  cols_for_meta,
  cores
)
}
\arguments{
\item{object}{S4 object of class metab_analyser}

\item{which_data}{dataset or datasets to be used}

\item{lag}{which lagged model to use. 1 means one-lagged model, similary 2,3,..etc}

\item{timepoints}{timepoints of interest that are to be used to build networks(in the order of measurement)}

\item{alpha}{parameter for regression coefficient}

\item{nfolds}{nfolds parameter for glmnet style of regression}

\item{cols_for_meta}{a list of character vectors of column names to be used for visualization of the networks.}

\item{cores}{Number of cores to be used for the process}
}
\value{
temporal network data with edgelist and regression values
}
\description{
calculates temporal networks for each dataset with a lagged model as used in graphical VAR
}
