% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_helpers.R
\name{calc_time_trend}
\alias{calc_time_trend}
\title{Calculate feature time trends}
\usage{
calc_time_trend(object, which_data, name = NULL)
}
\arguments{
\item{object}{a S4 object of class "metime_analyser".}

\item{which_data}{dataset name to use.}

\item{name}{name for the results entry.}
}
\value{
A metime_analyser object with results appended.
}
\description{
Fits a linear trend per feature over time.
}
