% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_calc.R
\name{calc_trajectories_by_mean}
\alias{calc_trajectories_by_mean}
\title{Function to get mean trajectories of metabolites and phenotypic traits}
\usage{
calc_trajectories_by_mean(object, which_data, columns)
}
\arguments{
\item{object}{An S4 object of class metime_analyser}

\item{which_data}{Dataset of interest}

\item{columns}{Other data that you want to see along with metabolites(column names from rowdata)}
}
\value{
plot_data table that can be used to make the plotter object without metadata
}
\description{
function to extract mean trajectories
}
