% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_calc.R
\name{calc_ttest}
\alias{calc_ttest}
\title{Function to calculate students t-test}
\usage{
calc_ttest(object, which_data, split_var)
}
\arguments{
\item{object}{S4 object of class metime_analyser}

\item{which_data}{dataset or datasets to be used for the analysis}

\item{split_var}{split variable for testing such as diagnostic group etc}

\item{timepoints}{two timepoints of interest to perform the test on}
}
\value{
t-test result as a list or a list of t-test results
}
\description{
Method for S4 object of class metime_analyser for performing t-test
}
