% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_calc.R
\name{calc_ttest_samples}
\alias{calc_ttest_samples}
\title{Function to calculate students t-test between samples at different timepoints}
\usage{
calc_ttest_samples(object, which_data, timepoints, type, paired)
}
\arguments{
\item{object}{S4 object of class metime_analyser}

\item{which_data}{dataset or datasets to be used for the analysis}

\item{timepoints}{timepoints of interest to perform the test on}

\item{type}{type of ttest to be used either "two.sided", "less", or "greater"}

\item{paired}{Logical to perform paired t.test or not}
}
\value{
plotter object with t-test results
}
\description{
Method for S4 object of class metime_analyser for performing t-test
}
