% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_col_normality.R
\name{check_col_normality}
\alias{check_col_normality}
\title{Function to check for col_normality data whether it is added or not.}
\usage{
check_col_normality(object, which_data)
}
\arguments{
\item{object}{S4 object of class of metime_analyser}

\item{which_data}{dataset/s to check}
}
\value{
NULL if it passes all the sanity checks
}
\description{
function to check whether col_normality data is added to
the object or not
}
