% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ids_and_classes.R
\name{check_ids_and_classes}
\alias{check_ids_and_classes}
\title{Function to check the ids in the data and data format}
\usage{
check_ids_and_classes(object)
}
\arguments{
\item{object}{S4 object of class of metime_analyser}
}
\value{
NULL if it passes all the sanity checks
}
\description{
sanity check to check for ids and order of the data
}
