% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{check_results}
\alias{check_results}
\title{Function to check the format of results if they exist}
\usage{
check_results(object)
}
\arguments{
\item{object}{S4 object of class of metime_analyser}
}
\value{
NULL if it passes all the sanity checks
}
\description{
sanity check to check for results of the analysis
}
