% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{check_rownames_and_colnames}
\alias{check_rownames_and_colnames}
\title{Function to check the format of rownames and colnames and if they are same or not}
\usage{
check_rownames_and_colnames(object)
}
\arguments{
\item{object}{S4 object of class of metime_analyser}
}
\value{
NULL if it passes all the sanity checks
}
\description{
sanity check to check for rownames of the data
}
