% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_check.R
\name{check_scaling_and_transformation}
\alias{check_scaling_and_transformation}
\title{Function to check if the data is already scaled or log transformed}
\usage{
check_scaling_and_transformation(object, which_data)
}
\arguments{
\item{object}{An S4 object of metime_anlyser class}

\item{which_data}{the dataset/s to be checked}
}
\value{
NULL but checks if the data is scaled or not
}
\description{
Function to be applied on metime_analyser to check for log transformation and scaling
}
