% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_comp.R
\name{comp_network_with_established}
\alias{comp_network_with_established}
\title{Function to compare the network generated from data to an existing network}
\usage{
comp_network_with_established(calc_networks, est_network)
}
\arguments{
\item{calc_networks}{list of networks calculated from data}

\item{est_network}{established network to compare with the established network results
Make sure that this network has only two columns with names as node1 and node2}
}
\value{
fisher test results with pval and test statistic
}
\description{
function to perform fishers exact test to decide which network is the best.
}
