% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_regression.R
\name{comp_regression}
\alias{comp_regression}
\title{Comparison of regression analysis}
\usage{
comp_regression(object, result_index = NULL, method = "sign")
}
\arguments{
\item{object}{a S4 object of class metime_analyser.}

\item{result_index}{a character vector of results to compare.}

\item{method}{a character vector of methods to be applied for comparison. Options: 'sign', 'heterogeneity'. See details for more information.}
}
\value{
a S4 object of class metime_analyser object with function output appended to results
}
\description{
A function to compare results from regression analysis (LM, LMM, GAMM)
}
\details{
Multiple options are available to compare betas and betas + se from regression analysis. Sign: comparison of beta signs. heterogeneity: comparison of betas and se.
}
