% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_general.R
\name{convert_s4_to_s3}
\alias{convert_s4_to_s3}
\title{Function to Convert S4 object of class metab_analyser to an S3 object with same architecture}
\usage{
convert_s4_to_s3(object)
}
\arguments{
\item{object}{An object of class metab_analyser}
}
\value{
An S3 object of the same data as metab_analyser in other words all slots are now converted into nested lists
}
\description{
Function to Convert S4 object of class metab_analyser to an S3 object with same architecture
}
