% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_study_characteristics.R
\name{distribution_plotter}
\alias{distribution_plotter}
\title{Function for Plotting distributions of phenotypic variables}
\usage{
distribution_plotter(object, colname, which_data)
}
\arguments{
\item{object}{An object of class metab_analyser}

\item{colname}{Name of the variable whose distribution is of interest}

\item{which_data}{Name of the dataset from which the samples will be extracted}
}
\value{
a list with either 1) density plot, mean table acc to timepoint and variable type or
2) bar plot, line plot, and variable type
}
\description{
Function for Plotting distributions of phenotypic variables
}
