% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_append_analyser_object.R
\name{get_append_analyser_object}
\alias{get_append_analyser_object}
\title{This function appends an object of class metime_analyser with a new dataset.}
\usage{
get_append_analyser_object(object, data, col_data, row_data, name)
}
\arguments{
\item{object}{S4 object of class metime_analyser}

\item{data}{data.frame containing data}

\item{col_data}{data.frame containing col_data: id column of col data has to match colnames of data}

\item{row_data}{data.frame containing row_data: id column of row data has to match rownames of data}

\item{name}{Name of the new dataset}
}
\value{
An object of class metime_analyser
}
\description{
function to apply on metime_analyse object to append a new dataset into the existing object
}
\examples{
# append data frames into the metime_analyser object
appended_object <- get_append_metab_object(object=metime_analyser_object, data=data, row_data=data, col_data=col_data, name="name of the new dataset")
}
