% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get.R
\name{get_append_metab_object}
\alias{get_append_metab_object}
\title{This function appends an object for MetabAnalyze with a new dataset.}
\usage{
get_append_metab_object(object, data, col_data, row_data, name = NULL)
}
\arguments{
\item{object}{S4 MetabAnalyze object}

\item{data}{data.frame containing row_data: id column of row data has to match rownames of data}
}
\value{
An object of class metab_analyser
}
\description{
This function appends an object for MetabAnalyze with a new dataset.
}
