% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_ggms.R
\name{get_betas}
\alias{get_betas}
\title{Function to perform regression on list of matrics either divided based on time(temporal net) or data type(Multibipartite Lasso)}
\usage{
get_betas(list_of_mats, alpha, nfolds, type_of_data)
}
\arguments{
\item{list_of_mats}{list of matrices that are divided based on platform or timepoints}

\item{alpha}{parameter for glmnet alpha=1 represents ridge regression and alpha=0 represents lasso regression and anything in between results in a mixed penalty regression}

\item{nfolds}{nfolds parameter for glmnet}
}
\value{
a list with different combinations used and each combination is a nested list with regression result data
}
\description{
performs regression for both temporal and multibipartite GGMs
}
