% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get.R
\name{get_betas_for_multibipartite_lasso}
\alias{get_betas_for_multibipartite_lasso}
\title{Function to perform multibipartite style regression on a list of matrices}
\usage{
get_betas_for_multibipartite_lasso(list_of_mats, alpha, nfolds)
}
\arguments{
\item{list_of_mats}{a list with matrices and samples ordered similarly}

\item{alpha}{alpha for cv.glmnet regression. Defines style of penalty.}

\item{nfolds}{nfolds for cv.glmnet}
}
\value{
returns a list with information of the combinations in context
}
\description{
Performs multibipartite lasso in cv.glmnet style on a list of matrices that have
metabolite information from different platforms
}
