% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coldata.R
\name{get_coldata}
\alias{get_coldata}
\title{Get col_data from a S4 object of the class "metime_analyser"}
\usage{
get_coldata(object, which_data)
}
\arguments{
\item{object}{a S4 object of the class "metime_analyser"}

\item{which_data}{a character of the dataset name of interest.}
}
\value{
A dataframe containing the col data of the dataset of interest.
}
\description{
Get the column information of a specified dataset from a S4 object of the class "metime_analyser"
}
