% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_common_samples_at_timepoints.R
\name{get_common_samples_at_timepoints}
\alias{get_common_samples_at_timepoints}
\title{Get common samples at multiple timepoints chosen}
\usage{
get_common_samples_at_timepoints(object, which_data, timepoints)
}
\arguments{
\item{which_data}{a character to define which dataset is to be used.}

\item{timepoints}{a character vector to define time points}

\item{objecta}{S4 object of the class "metime_analyzer".}
}
\value{
A character vector of common subjects at common time points.
}
\description{
Get a vector of common samples that can be used for stratification.
}
