% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get data from a S4 object of the class "metime_analyser"}
\usage{
get_data(object, which_data)
}
\arguments{
\item{object}{a S4 object of the class "metime_analyser"}

\item{which_data}{a character of the dataset name of interest.}
}
\value{
A dataframe containing the data of the dataset of interest.
}
\description{
Get the data of a specified dataset from a S4 object of the class "metime_analyser"
}
