% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get.R
\name{get_data_as_list}
\alias{get_data_as_list}
\title{This function that extracts data from the MetabAnalyze object as a list}
\usage{
get_data_as_list(object, which_data)
}
\arguments{
\item{object}{S4 MetabAnalyze object}

\item{which_data}{which data should we extract}
}
\value{
list of data
}
\description{
This function that extracts data from the MetabAnalyze object as a list
}
