% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files_and_names.R
\name{get_files_and_names}
\alias{get_files_and_names}
\title{Pack a dataset (data, row_data, col_data) into a single object of class "metime_analyser".}
\usage{
get_files_and_names(path, annotations_index)
}
\arguments{
\item{path}{a character defining the path to the parent directory}

\item{annotations_index}{a named list to be filled as
list(phenotype="Name or index of the files",
medication="Name or index of the files")}
}
\value{
An object of class "metime_analyser".
}
\description{
This function loads all the files from the parent directory. It assumes a
certain naming pattern as follows: "datatype_None|col|row_data.rds"
Any other naming pattern is not allowed. The function first writes
all files into a list and each type of data is packed into its respective
class i.e. col_data, row_data or data.
}
\examples{
get_files_and_names(path="/path/to/parent/directory", 
	annotations_index=list(phenotype="Name of phenotype file", 
medication="name of phenotype file"))
}
