% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get.R
\name{get_ggm_genenet}
\alias{get_ggm_genenet}
\title{Function to calculate a dynamic GeneNet GGM from a longitudnal data matrix}
\usage{
get_ggm_genenet(data, threshold = c("bonferoni", "FDR", "li"))
}
\arguments{
\item{data}{data matrix in a longitudnal format}

\item{threshold}{type of multiple hypothesis correction. Available are Bonferoni("bonferoni"),
Benjamini-Hochberg("FDR") and independent tests method("li", also see Li et al ....)}
}
\value{
a dataframe with edges, partial correlation and associated p-values
}
\description{
calculates GGM on longitudnal data matrix and returns a dataframe with edges,
partial correlation and associated p-values
}
