% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get.R
\name{get_make_metab_object}
\alias{get_make_metab_object}
\title{Function to pack all the data into a single object of class "metime_analyser"}
\usage{
get_make_metab_object(data, col_data, row_data, annotations_index, name = NULL)
}
\arguments{
\item{data}{data.frame containing data}

\item{col_data}{data.frame containing col_data: id column of col data has to match colnames of data}

\item{row_data}{data.frame containing row_data: id column of row data has to match rownames of data}

\item{annotations_index}{a list to be filled as follows = list(phenotype="Name or index of the file/list", medication="Name or index of the files/list")}

\item{name}{character. Name you want to assign to the new dataset that is being added on}
}
\value{
An object of class metime_analyser
}
\description{
This function creates an object of class metime_analyser from a dataset.
}
\examples{
# new_metime_analyser_object <- get_make_metab_object(data=data_frame, col_data=col_data_frame, row_data=row_data, name="name of the new dataset", 
                               annotations_index=list(phenotype="name of phenotype", medication="name of medication"))
}
