% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get.R
\name{get_make_plotter_object}
\alias{get_make_plotter_object}
\title{Function to make a plottable object for viz functions}
\usage{
get_make_plotter_object(data_list, metadata_list, plot_type, aesthetics)
}
\arguments{
\item{data_list}{list of plotable data}

\item{metadata_list}{list of metadata for each plot in data list}

\item{plot_type}{type of the plot you want to build. eg: "box", "dot" etc}

\item{aesthetics}{aesthetics for the plot object}
}
\description{
function to generate metime_plotter object from plot data and metadata
}
