% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_make_results.R
\name{get_make_results}
\alias{get_make_results}
\title{Get results list for S4 object of class "metime_analyser" object}
\usage{
get_make_results(object, data, metadata, calc_type, calc_info, name)
}
\arguments{
\item{object}{a S4 object of the class "metime_analyzer".}

\item{data}{a list of dataframes of plotable data obtained from any calc function.}

\item{metadata}{a dataframe or a list of dataframes with the metadata for the plot table mentioned above. To obtain these see
get_metadata_for_rows() and get_metadata_for_columns().}

\item{calc_type}{a character vector to specify type of calculation. Should be the same length as the list data provided.}

\item{calc_info}{a character to define the information about calculation, should be the same length as the list data provided.}

\item{name}{a character to be used as the index of the result.}
}
\value{
A S4 object of class "metime_analyser" with the result appended to the list of results.
}
\description{
Compile a result element for metime_analyser object.
}
