% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get.R
\name{get_metadata_for_mean_trajectories}
\alias{get_metadata_for_mean_trajectories}
\title{Function to get metadata for mean trajectories}
\usage{
get_metadata_for_mean_trajectories(object, columns)
}
\arguments{
\item{object}{an S4 metime analyser object}

\item{columns}{columns of interest for which we need the metadata for}
}
\value{
a metadata dataframe that will be used for plotting
}
\description{
Function to be applied on the metime analyser object to extract
metadata for the mean trajectories plot
}
