% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_general.R
\name{get_metadata_for_plotting}
\alias{get_metadata_for_plotting}
\title{Function to extract metadata of the metabolites}
\usage{
get_metadata_for_plotting(object, which_data, metab_groups, metab_ids)
}
\arguments{
\item{object}{S4 object of class metab_analyse}

\item{which_data}{choose the dataset from which metabolites will be extracted for metadata}

\item{metab_groups}{choose the column that has metabolite groups}

\item{metab_ids}{chodse the column that has metabolite names}
}
\value{
metadata dataframe with names, groups and class
}
\description{
Function to extract metadata of the metabolites
}
