% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata_for_rows.R
\name{get_metadata_for_rows}
\alias{get_metadata_for_rows}
\title{Get metadata for rows(in most cases for samples)}
\usage{
get_metadata_for_rows(object, which_data, columns)
}
\arguments{
\item{object}{a S4 object of the class "metime_analyzer".}

\item{which_data}{a character to define which dataset is to be used.}

\item{columns}{A list of character vectors for the columns of interest. Length of the list should be same as length of which_data.}
}
\value{
A data frame with metadata information for rows.
}
\description{
Generate a list of metadata for the S4 object of the class "metime_analyzer".
}
