% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get.R
\name{get_palette}
\alias{get_palette}
\title{Get a palette of "n" distinct colorblind friendly colors}
\usage{
get_palette(n)
}
\arguments{
\item{n}{number of colors wanted in the palette}
}
\value{
a color palette vector with colors in the form of hex codes
}
\description{
Function to get a palette of distinct colorblind friendly colors, the distinctiveness is determined by the difference in their hue values.
}
