% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters_of_results.R
\name{get_parameters_of_results}
\alias{get_parameters_of_results}
\title{Get parameters of the results}
\usage{
get_parameters_of_results(object, index)
}
\arguments{
\item{object}{An S4 object of class metime_analyser.}

\item{index}{a character or numeric value to be used as the result index.}
}
\value{
a dataframe with functions(rownames) and parameters(colnames).
}
\description{
Get parameters and functions applied to obtain results.
}
