% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_report.R
\name{get_report,metime_analyser-method}
\alias{get_report,metime_analyser-method}
\title{\n',paste0("Calculation type:  ", results[\link{section_name}][\link{"information"}]$calc_type)\%>\% gsub(pattern='"', replacement = "'", fixed=T),}
\usage{
\S4method{get_report}{metime_analyser}(
  object,
  title = NULL,
  file = NULL,
  table = F,
  device = "html",
  interactive = F,
  author = NULL
)
}
\description{
\n',paste0("Calculation type:  ", results[\link{section_name}][\link{"information"}]$calc_type)\%>\% gsub(pattern='"', replacement = "'", fixed=T),
}
