% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{get_results}
\alias{get_results}
\title{Get results element by index}
\usage{
get_results(object, index)
}
\arguments{
\item{object}{a S4 object of the class "metime_analyzer".}

\item{index}{a character or numeric value defining which result should be}
}
\value{
A list with result elements that
the functions applied and type of calculation
}
\description{
Get results from S4 object of class metime_analyser by a specified index. Print object in console to see all available indices.
}
