% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get.R
\name{get_rowdata}
\alias{get_rowdata}
\title{Function to extract row data of a dataset}
\usage{
get_rowdata(object, which_data)
}
\arguments{
\item{object}{An object of class S4}

\item{which_data}{Dataset of interest}
}
\value{
row data of the dataset of interest
}
\description{
Function to get rowdata
}
