% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_importer.R
\name{get_stratified_data}
\alias{get_stratified_data}
\title{Function to stratify before calculation}
\usage{
get_stratified_data(object, which_data, stratifications)
}
\arguments{
\item{object}{An S4 object of class metime_analyser}

\item{which_data}{Dataset to be used}

\item{stratifications}{list of variables and their values to stratified - list(name=value)}
}
\value{
data with stratifications aforementioned
}
\description{
Stratification of a dataset in calculations. Used only in calc_ functions
}
