% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_general.R
\name{get_text}
\alias{get_text}
\title{Function to Obtain textual information for visualization in interactive plots}
\usage{
get_text(data, colnames)
}
\arguments{
\item{data}{a dataframe with plotting data along with other variables for visualization}

\item{colnames}{a character vector with the names of the variables that you want to see on the plot}
}
\value{
a vector with strings that can be parsed into plot_ly text.
}
\description{
Function to Obtain textual information for visualization in interactive plots
}
