% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get.R
\name{get_text_for_plot}
\alias{get_text_for_plot}
\title{Function to Obtain textual information for visualization in interactive plots}
\usage{
get_text_for_plot(data, colnames)
}
\arguments{
\item{data}{a dataframe with plotting data along with other variables for visualization}

\item{colnames}{a character vector with the names of the variables that you want to see on the plot}
}
\value{
a vector with strings that can be parsed into plot_ly text.
}
\description{
a standard function to be applied on data matrices or dataframes with the colnames of interest such that the information from
columns is visualized in the interactive plot
}
\examples{
# text = get_text(data=data.frame, colnames=c("names","of","columns", "of", "interest"))
}
