% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_ggms.R
\name{ggm_visualizer}
\alias{ggm_visualizer}
\title{Function to plot data from network and object after calculating a certain ggm}
\usage{
ggm_visualizer(network, type, metadata)
}
\arguments{
\item{network}{dataframe with information of the network}

\item{type}{Type of visualization options = c("visNetwork", "Cytoscape", "networkD3")}

\item{metadata}{dataframe with two columns 1) metabs and 2) Group they belong to}
}
\value{
network plot based on the type chosen by the user
}
\description{
A function to plot ggms of different kinds namely visNetwork, cytoscape and networkD3
}
