% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_conservation.R
\name{meta_conservation}
\alias{meta_conservation}
\title{Meta comparison for conservation index results}
\usage{
meta_conservation(
  object,
  result_index = NULL,
  top_k = 50,
  name = "meta_conservation_1"
)
}
\arguments{
\item{object}{a S4 object of class metime_analyser or a list of two metime_analyser objects}

\item{result_index}{character/numeric input for results. If NULL, all matching results are used.}

\item{top_k}{numeric indicating the top-K features used for overlap calculations}

\item{name}{a character input to set the name of the results}
}
\value{
An S4 object of class meta_results with the compared results and meta results
}
\description{
Compare conservation index results within a result or across results.
}
