% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_feature_overlap.R
\name{meta_feature_overlap}
\alias{meta_feature_overlap}
\title{Meta comparison for feature overlap}
\usage{
meta_feature_overlap(
  object,
  result_index = NULL,
  name = "meta_feature_overlap_1"
)
}
\arguments{
\item{object}{a S4 object of class metime_analyser or a list of two metime_analyser objects}

\item{result_index}{character/numeric input for results. If NULL, all matching results are used.}

\item{name}{a character input to set the name of the results}
}
\value{
An S4 object of class meta_results with the compared results and meta results
}
\description{
Compare feature selection outputs within or across results.
}
