% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mapping.R
\name{meta_fisher_for_networks}
\alias{meta_fisher_for_networks}
\title{Function to map edge lists to HMDB and KEGG IDs}
\usage{
meta_fisher_for_networks(object, results_index, metanno, app_lookup)
}
\arguments{
\item{object}{An S4 object of class metime_analyser}

\item{results_index}{Name/index for the network results}

\item{metanno}{Dataframe with mappings}
}
\value{
mapped results
}
\description{
Function to map metabolite names
}
