% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_helpers.R, R/meta_matrix_similarity.R
\name{meta_matrix_similarity}
\alias{meta_matrix_similarity}
\title{Meta comparison for matrix similarity}
\usage{
meta_matrix_similarity(
  object,
  result_index = NULL,
  name = "meta_matrix_similarity_1"
)

meta_matrix_similarity(
  object,
  result_index = NULL,
  name = "meta_matrix_similarity_1"
)
}
\arguments{
\item{object}{a S4 object of class metime_analyser or a list of two metime_analyser objects}

\item{result_index}{character/numeric input for results. If NULL, all matching results are used.}

\item{name}{a character input to set the name of the results}
}
\value{
An S4 object of class meta_results with the compared results and meta results

An S4 object of class meta_results with the compared results and meta results
}
\description{
Compare pairwise distance or correlation results within or across results.

Compare pairwise distance or correlation results within or across results.
}
