% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_network.R
\name{meta_network}
\alias{meta_network}
\title{comparison of network analysis}
\usage{
meta_network(object, results_index, method, name = "meta_network_1")
}
\arguments{
\item{object}{An S4 object of class metime_analyser}

\item{results_index}{character vector or numeric vector of length at least 2 which contains indices of results of interest}

\item{method}{character vector to define methods of interest. Accepted inputs are: "sign" for sign test and "het" for
heterogenity test}

\item{name}{a character containing the name of the new result element appended to the analyzer object.
Default set to 'meta_network_1'.}
}
\value{
object with meta analyses of networks appended to results
}
\description{
A function to compare results from network analysis (GGMs)
}
\seealso{
\link{meta_parametric_test}, \link{meta_nonparametric_test}, \link{meta_regression}
}
