% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_nonparametric_test.R
\name{meta_nonparametric_test}
\alias{meta_nonparametric_test}
\title{Function to perform non-parametric tests on two sets of results or results of interest}
\usage{
meta_nonparametric_test(
  object,
  results_index,
  type,
  formula,
  name = "meta_nonparametric_test_1",
  ...
)
}
\arguments{
\item{object}{An S4 object of class metime_analyser}

\item{results_index}{A character/numeric input for a results. The length should be equal to 1.}

\item{type}{character input to define the type of test. Two types are available: "kruskal", "wilcox"}

\item{formula}{formula(class) input for the test functions. See rstatix::wilcox_test() or rstatix::kruskal_test()}

\item{name}{character input to set the name of the results}

\item{...}{additional arguments for either rstatix::wilcox_test() or rstatix::kruskal_test() based on input}
}
\value{
An S4 object of class metime_analyser with this meta analysis stored as results
}
\description{
Meta function to perform non-parametric tests on different types of results to perform meta analysis of interest
}
\seealso{
\link{meta_parametric_test} \link[rstatix:wilcox_test]{rstatix::wilcox_test} \link[rstatix:kruskal_test]{rstatix::kruskal_test}
}
