% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_parametric_test.R
\name{meta_parametric_test}
\alias{meta_parametric_test}
\title{Function to perform parametric tests on two sets of results or results of interest}
\usage{
meta_parametric_test(
  object,
  results_index,
  type,
  formula,
  name = "meta_parametric_test_1",
  ...
)
}
\arguments{
\item{object}{An S4 object of class metime_analyser}

\item{results_index}{A character/numeric input for a results. The length should be equal to 1.}

\item{type}{character input to define the type of test. Two types are available: "t_test", "anova"}

\item{formula}{formula(class) input for the test functions. See rstatix::t_test() or rstatix::anova_test()}

\item{name}{character input to set the name of the results}

\item{...}{additional arguments for either rstatix::t_test() or rstatix::anova_test() based on input}
}
\value{
An S4 object of class metime_analyser with this meta analysis stored as results
}
\description{
Meta function to perform parametric tests on different types of results to perform meta analysis of interest
}
\seealso{
\link{meta_nonparametric_test}, \link[rstatix:t_test]{rstatix::t_test}, \link[rstatix:anova_test]{rstatix::anova_test}
}
