% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_regression.R
\name{meta_regression}
\alias{meta_regression}
\title{Meta comparison for regression outputs}
\usage{
meta_regression(
  object,
  method = c("sign", "cor", "het"),
  result_index = NULL,
  name = "meta_regression_1"
)
}
\arguments{
\item{object}{a S4 object of class metime_analyser or a list of two metime_analyser objects}

\item{method}{a character vector of methods 'sign', 'cor', 'het'}

\item{result_index}{character/numeric input for results. If NULL, all matching results are used.}

\item{name}{a character input to set the name of the results}
}
\value{
An S4 object of class meta_results with the compared results and meta results
}
\description{
Compare regression outputs within or across results.
}
