% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get.R, R/fun_viz.R
\docType{class}
\name{metab_plotter-class}
\alias{metab_plotter-class}
\title{add aesthetics to plot and so on in this style
lol <- aes(x=x, y=y)
empty_plot$mappings <- lol
creating metab_plotter class that converts calculations and metadata as a plotable object to parse into viz_dot_plotter, viz_heatmap_plotter etc
Contains slots - plot_data: Dataframe with plotting data and metadata for visualization
- plot_parameters: ggplot() object with predefined aesthetics
- aesthetics: list to define aesthetics. Eg: aesthetics=list(x="colname.x", y="colname.y", color="color", shape="shape")
- the example above will be predefined in all the methods that creates this object.}
\description{
add aesthetics to plot and so on in this style
lol <- aes(x=x, y=y)
empty_plot$mappings <- lol
creating metab_plotter class that converts calculations and metadata as a plotable object to parse into viz_dot_plotter, viz_heatmap_plotter etc
Contains slots - plot_data: Dataframe with plotting data and metadata for visualization
- plot_parameters: ggplot() object with predefined aesthetics
- aesthetics: list to define aesthetics. Eg: aesthetics=list(x="colname.x", y="colname.y", color="color", shape="shape")
- the example above will be predefined in all the methods that creates this object.

creating metab_plotter class that converts calculations and metadata as a plotable object to parse into viz_dot_plotter, viz_heatmap_plotter etc
Contains slots - plot_data: Dataframe with plotting data and metadata for visualization
- plot_parameters: ggplot() object with predefined aesthetics
- aesthetics: list to define aesthetics. Eg: aesthetics=list(x="colname.x", y="colname.y", color="color", shape="shape")
- the example above will be predefined in all the methods that creates this object.
}
