% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeTime-package.R
\docType{package}
\name{MeTime-package}
\alias{MeTime}
\alias{MeTime-package}
\title{MeTime: Metabolomics Time}
\description{
R Package to aid in longitudinal data analysis with a focus on metabolomics data.
}
\author{
\strong{Maintainer}: Bharadwaj Marella \email{bharadwaj.marella@helmholtz-muenchen.de}

Authors:
\itemize{
  \item Patrick Weinisch \email{patrick.weinisch@helmholtz-muenchen.de}
  \item Matthias Arnold \email{matthias.arnold@helmholtz-muenchen.de}
}

}
\keyword{internal}
