% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_mod.R
\name{mod_average_combine}
\alias{mod_average_combine}
\title{Function to combine replicates in the data}
\usage{
mod_average_combine(object, which_data)
}
\arguments{
\item{object}{An S4 object of class metime_analyser}

\item{which_data}{The dataset in which the data points are to be combined}
}
\value{
object with combined information
}
\description{
Function to combine two data points of a sample in a dataset
}
