% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_mod.R
\name{mod_code_metab_names}
\alias{mod_code_metab_names}
\title{Function to convert metabolite names to IDs}
\usage{
mod_code_metab_names(object, which_data, table)
}
\arguments{
\item{object}{An S4 object of class metime_analyser}

\item{which_data}{character vector to define the datasets to use}

\item{table}{logical vector. If TRUE returns table of names to IDs for mapping}
}
\value{
A list with S4 object and list of mapping tables, the object can be used for GGMs
}
\description{
Function to convert metabolite names to IDs
}
