% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_mod.R
\name{mod_convert_s4_to_s3}
\alias{mod_convert_s4_to_s3}
\title{Function to Convert S4 object of class metime_analyser to an S3 object with same architecture}
\usage{
mod_convert_s4_to_s3(object)
}
\arguments{
\item{object}{An object of class metime_analyser}
}
\value{
An S3 object of the same data as metime_analyser in other words all slots are now converted into nested lists
}
\description{
converter function to be applied onto metime_analyse object to convert into a standard list of S3 type.
}
\examples{
# convert S4 object to a list
s3_list <- mod_convert_s4_to_s3(object=metime_analyser_object)
}
