% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_extract_common_samples.R
\name{mod_extract_common_samples}
\alias{mod_extract_common_samples}
\title{Function to extract common samples across all datasets and store them only}
\usage{
mod_extract_common_samples(object)
}
\arguments{
\item{object}{An object of class metime_anaylser}
}
\value{
metime_analyser object with only common samples across all datasets present in the object parsed
}
\description{
Modification(mod) function applied on object of class metime_analyser to extract common samples across datasets.
}
\examples{
# extracting common samples across all datasets
new_object_with_only_common_samples <- mod_extract_common_samples(object=metime_analyser_object)
}
