% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filter.R
\name{mod_filter}
\alias{mod_filter}
\title{Modification function to filter columns in data, row_data or col_data}
\usage{
mod_filter(object, which_data, ..., type = "data")
}
\arguments{
\item{object}{A S4 object of class metime_analyser.}

\item{which_data}{a vector of character defining which data/result is to be filtered.}

\item{...}{arguments to pass directly into dplyr::filter() function.}

\item{type}{character input of length 1 to define the type of data to be manipulated. Accepted inputs are "row_data", "col_data", "data"
and "results". However renamed results will be returned to the user as a list of results and will not return the full analyser
object}
}
\value{
object with mutated data, col_data or row_data. However, if type is "results" then plot_data with modifications expected
will be returned
}
\description{
Modification (mod) function to filter columns in data, row_data or col_data. It is a wrapper function to dplyr::filter()
made for metime_analyser objects.
}
\seealso{
\link{mod_mutate}, \link{mod_rename}
}
