% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_helpers.R
\name{mod_filter_features_by_missingness}
\alias{mod_filter_features_by_missingness}
\title{Filter features by missingness}
\usage{
mod_filter_features_by_missingness(object, threshold = 0.2, which_data = NULL)
}
\arguments{
\item{object}{a S4 object of class "metime_analyser".}

\item{threshold}{numeric between 0 and 1 defining maximum allowed missingness.}

\item{which_data}{character vector of dataset names to filter. If NULL, filters all datasets.}
}
\value{
A modified metime_analyser object.
}
\description{
Removes features with missingness above a threshold.
}
