% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_helpers.R
\name{mod_filter_features_by_variance}
\alias{mod_filter_features_by_variance}
\title{Filter features by variance}
\usage{
mod_filter_features_by_variance(object, min_variance = 0, which_data = NULL)
}
\arguments{
\item{object}{a S4 object of class "metime_analyser".}

\item{min_variance}{numeric defining minimum variance required to keep a feature.}

\item{which_data}{character vector of dataset names to filter. If NULL, filters all datasets.}
}
\value{
A modified metime_analyser object.
}
\description{
Removes features with variance below a threshold.
}
