% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_helpers.R
\name{mod_filter_timepoints}
\alias{mod_filter_timepoints}
\title{Filter timepoints from datasets}
\usage{
mod_filter_timepoints(object, timepoints, which_data = NULL)
}
\arguments{
\item{object}{a S4 object of class "metime_analyser".}

\item{timepoints}{character or numeric vector of timepoints to keep.}

\item{which_data}{character vector of dataset names to filter. If NULL, filters all datasets.}
}
\value{
A modified metime_analyser object.
}
\description{
Removes samples that are not in the specified timepoints.
}
