% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_mod.R
\name{mod_logtrans}
\alias{mod_logtrans}
\title{Function to apply log transformation}
\usage{
mod_logtrans(object, which_data, base)
}
\arguments{
\item{object}{An object of class metime_analyser}

\item{which_data}{Name of the dataset to be used}

\item{base}{base of log to be used}
}
\value{
An object of class metime_analyser with processed data
}
\description{
Function to log transform data
}
\examples{
# example to apply log transformation
object <- mod_logtrans(object, which_data="name of the dataset", base=2)
}
