% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_merge_metime_analysers.R
\name{mod_merge_metime_analysers}
\alias{mod_merge_metime_analysers}
\title{Function to merge one or more metime_analyser objects}
\usage{
mod_merge_metime_analysers(list_of_objects, annotations_index)
}
\arguments{
\item{list_of_objects}{list of metime analyser objects that are to be merged. The objects must be named
Ex: list_of_objects <- list(object1=object1, object2=object2, ...). If not named then they will be named by the function}

\item{annotations_index}{new list with annotations_index. Can also set to be NULL.}
}
\value{
A merged metime_analyser object
}
\description{
function to merge multiple metime_analyser objects
}
