% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_merge_metime_plotters.R
\name{mod_merge_metime_plotters}
\alias{mod_merge_metime_plotters}
\title{Function to combine plotter objects}
\usage{
mod_merge_metime_plotters(list_of_objects, groups)
}
\arguments{
\item{list_of_objects}{list of plotter objects to be merged}

\item{groups}{character vector to define the groups that are involved}
}
\value{
plotter object merged with all the plotters
}
\description{
Function to combine plotter objects based on similar calc type
}
