% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_merge_results.R
\name{mod_merge_results}
\alias{mod_merge_results}
\title{Function to merge two sets of results}
\usage{
mod_merge_results(object, results_index, sub_results = 1, groups, name)
}
\arguments{
\item{object}{An S4 object of class metime_analyser}

\item{results_index}{character of numeric vector to define the index of results to merge.
if length of results_index is 1, then the plot_data list is merged into a single data.frame}

\item{sub_results}{Vector to define indices of plot_data to be merged. Set to 1 by default.}

\item{groups}{character vector to define the two sets of different results that are being merged}

\item{name}{character to name the new merged results}
}
\value{
metime_analyser object with merged results
}
\description{
Modification(mod) function to merge two sets of results based on calc_type
}
\seealso{
\link{mod_merge_data}
}
