% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mutate.R
\name{mod_mutate}
\alias{mod_mutate}
\title{Function to mutate columns in row_data or col_data}
\usage{
mod_mutate(object, which_data, type = "data", ...)
}
\arguments{
\item{object}{An S4 object of class metime_analyser}

\item{which_data}{character or numeric input to define Dataset/result of interest. Has to be of length=1}

\item{type}{character input of length 1 to define the type of data to be manipulated. Accepted inputs are "row_data", "col_data", "data"
and "results". However renamed results will be returned to the user as a list of results and will not return the full analyser
object}

\item{...}{arguments to pass directly into dplyr::mutate() function.}
}
\value{
object with mutated data
}
\description{
Modification(mod) function to mutate data, row_data, col_data and results. It is a wrapper function to dplyr::mutate()
made for metime_analyser objects.
}
\seealso{
\link{mod_filter}, \link{mod_rename}
}
