% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_mod.R
\name{mod_prep_data_for_ggms}
\alias{mod_prep_data_for_ggms}
\title{Function to prepare and preprocess S4 objects to use it for gaussian gaphical models. Also converts S4 to S3}
\usage{
mod_prep_data_for_ggms(object, which_type, mlp_or_temp)
}
\arguments{
\item{object}{An object of class metime_analyser}

\item{which_type}{two choices either: 1) single -  converts S4 to S3 and returns the nested list
2) multi - extracts common samples across the dataframes and returns an S3 nested list}

\item{mlp_or_temp}{boolean. If true preps data for multibipartite lasso or temporal networks}
}
\value{
An S3 object(nested list) with the same architecture as that of class metime_analyser
}
\description{
function to be applied onto metime_analyse object to convert into a standard list of S3 type based on the type of GGM analysis to be performed.
}
\examples{
# prepping data for genenet ggm for single dataset
object <- mod_prep_data_for_ggms(object, which_type="single", mlp_or_temp=FALSE)
}
