% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_remove_duplicates.R
\name{mod_remove_duplicates}
\alias{mod_remove_duplicates}
\title{Function to remove duplicates}
\usage{
mod_remove_duplicates(object)
}
\arguments{
\item{object}{An S4 object of class metime_analyser}
}
\value{
object after removing duplicated data
}
\description{
Function to remove duplicates from the analyser object
}
