% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_mod.R
\name{mod_remove_nas}
\alias{mod_remove_nas}
\title{Function to remove NA's from data matrices}
\usage{
mod_remove_nas(object, which_data)
}
\arguments{
\item{object}{S4 object of class metime_analyser}

\item{which_data}{dataset/s for which the method is to be applied}
}
\value{
S4 object with NA's removed and data manipulated accordingly
}
\description{
A method applied on S4 object to remove NA's and change data accordingly
}
