% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_rename.R
\name{mod_rename}
\alias{mod_rename}
\title{Function to rename columns in the datasets and results of analyser object}
\usage{
mod_rename(object, which_data, type = "data", ...)
}
\arguments{
\item{object}{An S4 object of class metime_analyser}

\item{which_data}{character or numeric input of length 1 to define the index of dataset or results}

\item{type}{character input of length 1 to define the type of data to be manipulated. Accepted inputs are "row_data", "col_data", "data"
and "results". However renamed results will be returned to the user as a list of results and will not return the full analyser
object}

\item{...}{arguments to pass to dplyr::rename to change the name of the functions. Example new_colname = "old_colname"}
}
\value{
metime_analyser object with mutated names of dataframes that are parsed
}
\description{
Function to rename columns of data, row_data, col_data and plot_data of the analyser object.
Arguments will be passed directly into dplyr::rename() function
}
\seealso{
\link{mod_filter} \link{mod_mutate}
}
