% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_mod.R
\name{mod_stratify_analyser}
\alias{mod_stratify_analyser}
\title{Function to stratify data in the metime analyser object}
\usage{
mod_stratify_analyser(object, which_data, variable)
}
\arguments{
\item{object}{S4 object of class metime_analyser}

\item{which_data}{Dataset/datasets to be used for stratification}

\item{variable}{Phenotype based on which the stratification would be performed}
}
\value{
list of metime_analyser objects which are stratified based on the variable chosen
}
\description{
Function to stratify the data of interest into different objects that can be used
to perform calculations according the said stratification variable
}
