% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_trans_foldchange.R
\name{mod_trans_foldchange}
\alias{mod_trans_foldchange}
\title{Function to calculate fold change}
\usage{
mod_trans_foldchange(object, which_data, log2 = T, col = "all")
}
\arguments{
\item{object}{an S4 object of class metime_analyser.}

\item{which_data}{a character vector defining the name of the dataset/s to be used.}

\item{log2}{logical input. TRUE denotes that the data is already log transformed.}

\item{col}{character vector to select the metabolites. Set to "all" for all metabolites.}
}
\value{
a S4 object of the class metime_analyzer with datasets updated with foldchange values.
}
\description{
Function to calculate relative log transformed concentration of
metabolites with respect to baseline concentration (foldchange) of multiple
datasets at once
}
