% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_trans_zscore.R
\name{mod_trans_zscore}
\alias{mod_trans_zscore}
\title{Function to scale the data}
\usage{
mod_trans_zscore(object, which_data)
}
\arguments{
\item{object}{An object of class metime_analyser}

\item{which_data}{character vector to define the dataset/s to be used}
}
\value{
An object of class metime_analyser with scaled which_data
}
\description{
Modification(mod) Function for scaling datasets
}
\examples{
# example to apply scaling
object <- mod_trans_zscore(object, which_data=c("dataset1", ...))
}
\seealso{
\link[base:scale]{base::scale}, \link{mod_trans_log}
}
