% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_study_characteristics.R
\name{pca_plotter_general}
\alias{pca_plotter_general}
\title{Function to Plot PCA for one dataset with samples as data points}
\usage{
pca_plotter_general(object, which_data, cols_for_vis)
}
\arguments{
\item{object}{An object of class metab_analyser}

\item{which_data}{Name of the dataset from which the samples will be extracted}

\item{cols_for_vis}{choose the colnames from phenotype data that you want to show in the text labels}
}
\value{
an interactive pca plot with text that can be modified.
}
\description{
Function to Plot PCA for one dataset with samples as data points
}
