% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_importer.R
\name{plot}
\alias{plot}
\title{Setting a plotting method for the metime_analyser class}
\usage{
plot(object, results_index, interactive, ...)
}
\arguments{
\item{object}{An S4 object of class metime_analyser}

\item{results_index}{Index/name of the results to be plotted}

\item{interactive}{logical. Set TRUE for interactive plot}

\item{...}{other parameters to pass color, fill, strat, viz(character vector with colnames for interactive)}
}
\value{
plots for a certain set of results
}
\description{
Function to plot results of a certain calculation
}
