% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_helpers.R
\name{plot_missingness}
\alias{plot_missingness}
\title{Plot missingness heatmap for a dataset}
\usage{
plot_missingness(object, which_data, max_features = 100)
}
\arguments{
\item{object}{a S4 object of class "metime_analyser".}

\item{which_data}{a single dataset name.}

\item{max_features}{maximum number of features to include in the plot.}
}
\value{
A ggplot object.
}
\description{
Creates a heatmap of missing values for a dataset.
}
