% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_general.R
\name{prep_data_for_ggms}
\alias{prep_data_for_ggms}
\title{Function to prepare and preprocess S4 objects to use it for gaussian gaphical models. Also converts S4 to S3}
\usage{
prep_data_for_ggms(object, which_type, mlp_or_temp)
}
\arguments{
\item{object}{An object of class metab_analyser}

\item{which_type}{two choices either: 1) single -  converts S4 to S3 and returns the nested list
2) multi - extracts common samples across the dataframes and returns an S3 nested list}
}
\value{
An S3 object(nested list) with the same architecture as that of class metab_analyser
}
\description{
Function to prepare and preprocess S4 objects to use it for gaussian gaphical models. Also converts S4 to S3
}
