% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_data.R
\name{rm_data}
\alias{rm_data}
\title{Function to remove datasets or results from the metime analyser object}
\usage{
rm_data(object, which_data, type = "dataset")
}
\arguments{
\item{object}{An S4 object of class metime_analyser}

\item{which_data}{index or name of the dataset to be removed or result}

\item{type}{"result" or "dataset" based on the type to be removed}
}
\value{
object with results/dataset removed
}
\description{
An S4 method to remove unwanted data
}
