% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_save.R
\name{save_analyser_object}
\alias{save_analyser_object}
\title{Function to extract analyser object data into a csv}
\usage{
save_analyser_object(object, which_data)
}
\arguments{
\item{object}{An object of class metime_plotter}

\item{which_data}{Character to specify the dataset}
}
\value{
saves the data in the working directory as a csv and returns nothing
}
\description{
extracts information from analyser object and saves it as a csv
}
\examples{
see examples here
save_analyser_object(object, which_data="dataset")
}
