% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot_from_plotter.R
\name{save_plot_from_plotter}
\alias{save_plot_from_plotter}
\title{Function to save interactive plots}
\usage{
save_plot_from_plotter(object, out)
}
\arguments{
\item{object}{An object of class metime_plotter}

\item{out}{Character to specify path of the output file to save the widget in}
}
\value{
saves the plot and returns nothing
}
\description{
extracts plot from plotter object and saves it as a widget
}
\examples{
save_plot_from_plotter(object)
}
