% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_save.R
\name{save_plotter_object}
\alias{save_plotter_object}
\title{Function to extract plot data into a csv}
\usage{
save_plotter_object(object, out)
}
\arguments{
\item{object}{An object of class metime_plotter}

\item{out}{Character to specify path of the output file or character vector in case of visNetwork}
}
\value{
saves the data into a csv and returns nothing
}
\description{
extracts information from plotter object and saves it as a csv
}
\examples{
see examples here
Network : save_plotter_object(object, out=c("edge.csv", "node.csv", "meta.csv"))
Others : save_plotter_object(object, out="outfile.csv")
}
