% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_parallel_cores.R
\name{set_parallel_cores}
\alias{set_parallel_cores}
\title{register parallel backend}
\usage{
set_parallel_cores(n_cores = NULL)
}
\arguments{
\item{n_cores}{A number of specified cores.}
}
\value{
set a parallel backend
}
\description{
function to run in order to perform the analysis parallely thereby saving time
}
