% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_importer.R
\name{show,metime_analyser-method}
\alias{show,metime_analyser-method}
\title{Setting new print definition for the metime_analyser object}
\usage{
\S4method{show}{metime_analyser}(object)
}
\arguments{
\item{object}{S4 object of class metime_analyser}
}
\value{
structure of the S4 object
}
\description{
function to see the structure of metime_analyser object
}
\examples{
structure(object)
}
