% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_importer.R
\name{show,metime_plotter-method}
\alias{show,metime_plotter-method}
\title{Setting new print definition for the metime_plotter object}
\usage{
\S4method{show}{metime_plotter}(object)
}
\arguments{
\item{object}{S4 object of class metime_plotter}
}
\value{
structure of the S4 object
}
\description{
function to see the structure of metime_plotter object
}
\examples{
structure(object)
}
