% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_general.R
\name{split_acc_time}
\alias{split_acc_time}
\title{Function to split the list of dataframes into a nested list with each dataframe
being split into into dataframes of different timepoints}
\usage{
split_acc_time(object)
}
\arguments{
\item{object}{An object of class metab_analyser}
}
\value{
list_of_data with each dataframe being broken into a list of dataframes with respect to the timepoint they belong to
}
\description{
Function to split the list of dataframes into a nested list with each dataframe
being split into into dataframes of different timepoints
}
