% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_general.R
\name{split_data_wrt_timepoints}
\alias{split_data_wrt_timepoints}
\title{Function to split data wrt to needed timepoints. Converts S4 back to S3 as well for other analysis.
Similar to split_acc_to_time however the timepoints can be manually chosen}
\usage{
split_data_wrt_timepoints(object, timepoints)
}
\arguments{
\item{object}{An object of class metab_analyser}

\item{timepoints}{A vector with timepoints of interest}
}
\value{
An S3 object(nested list) with the same architecture as that of class metab_analyser
}
\description{
Function to split data wrt to needed timepoints. Converts S4 back to S3 as well for other analysis.
Similar to split_acc_to_time however the timepoints can be manually chosen
}
