% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_importer.R
\name{structure}
\alias{structure}
\title{Setting new structure definition for the metime_analyser object}
\usage{
structure(object)
}
\arguments{
\item{object}{S4 object of class metime_analyser}
}
\value{
structure of the S4 object
}
\description{
function to see the structure of metime_analyser object
}
\examples{
structure(object)
}
