% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_study_characteristics.R
\name{tsne_plotter_general}
\alias{tsne_plotter_general}
\title{Function to plot tsne plots for one data set}
\usage{
tsne_plotter_general(object, which_data, metab_ids, metab_groups, cols_for_vis)
}
\arguments{
\item{object}{An object of class metab_analyser}

\item{which_data}{Name of the dataset from which the samples will be extracted}

\item{metab_ids}{name of the column with metabolite names in the col data matix}

\item{metab_groups}{name of the column which has the pathway information of the metabolites}

\item{cols_for_vis}{choose the colnames from phenotype data that you want to show in the text labels}
}
\value{
a list with two plot objects 1) samples - tSNE plot of the individuals(samples)
2) metabs - tSNE plot of the metabolites(metabs)
}
\description{
Function to plot tsne plots for one data set
}
