% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_study_characteristics.R
\name{tsne_plotter_wrt_common}
\alias{tsne_plotter_wrt_common}
\title{Function to plot tSNE plots for multiple datasets}
\usage{
tsne_plotter_wrt_common(
  object,
  which_data,
  metab_groups,
  metab_ids,
  cols_for_vis
)
}
\arguments{
\item{object}{An object of class metab_analyser}

\item{which_data}{a character vector - Names of the dataset from which the samples will be extracted}

\item{metab_groups}{a character vector - names of the column which has the pathway information of the metabolites(please make sure they are in the same order as the above ones)}

\item{metab_ids}{a character vector - names of the column with metabolite names in the col data matix(please make sure they are in the same order as the above ones)}

\item{cols_for_vis}{choose the colnames from phenotype data that you want to show in the text labels}

\item{phenotype_index}{index of the phenotype data. Can input either the name of the phenotype dataset or the index of the same}
}
\value{
a list with two plot objects 1) samples - tSNE plot of the individuals(".$samples")
2) metabs - tSNE plot of the metabolites(".$metabs")
}
\description{
Function to plot tSNE plots for multiple datasets
}
