% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_importer.R
\name{update_plots}
\alias{update_plots}
\title{Function to update plots post calculations}
\usage{
update_plots(object, .interactive = FALSE, type, results_index = NULL)
}
\arguments{
\item{object}{An S4 object of class metime_analyser}

\item{.interactive}{logical to make the plot interactive or not}

\item{type}{character to define the type of calculation used for updating the plot
cols_for_samples, cols_for_metabs, cols_for_meta etc will be used. So make sure you set those correctly
for better results.}

\item{results_index}{character/numeric input to define the results that you want to plot or replot with our automation.
Default will be set to NULL}
}
\value{
object with plots of the newest calculation
}
\description{
Function to update plots based on the calculation.
See a calc_* function to see the usage of this function
}
