% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_importer.R
\name{validity}
\alias{validity}
\title{Validity function to check if the object is valid or not}
\usage{
validity(object)
}
\arguments{
\item{object}{An S4 object of class metime_analyser}
}
\value{
logical suggesting if the object is intact or not
}
\description{
Function to check the validity of metime_analyser
}
\examples{
validObject(object)
}
