% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_viz.R
\name{viz_dimensionality_reduction}
\alias{viz_dimensionality_reduction}
\title{Function to dot plot any kind of dot_plotter including for dimensionality reduction}
\usage{
viz_dimensionality_reduction(
  data_list,
  metadata_list,
  axes_labels,
  title_metabs,
  title_samples
)
}
\arguments{
\item{data_list}{list obtained after applying calc_dimensionality_reduction() on metime_analyse object}

\item{metadata_list}{list obtained after applying get_metadata_for_plotting() on metime_analyse object}

\item{axes_labels}{character vector to specify the labels of the axes in the order x and y.}

\item{title_metabs}{character to specify the title of the plot of metabolites}
}
\value{
a list with both the plots of samples and metabolites. Can be accessed by using ".$samples" and ".$metabs"
}
\description{
General function to be implemented on data_list that is obtained after applying a dimensionality reduction method
}
