% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_viz.R
\name{viz_plotter}
\alias{viz_plotter}
\title{Setting up standard wrapper for all plot functions}
\usage{
viz_plotter(object, aesthetics)
}
\arguments{
\item{object}{S4 object of class metime_plotter}

\item{aesthetics}{list for aesthetics. eg: list(list(x="colname",y="colname",color="colname", shape="colname"), list(...)) for "dot" plot and "netowrk"
plot, for heatmap: list(x="colname", y="colname", fill="colname")}
}
\description{
plot function for metime_plotter object with different inputs to specialize plots. Used for all calc outputs.
}
