% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data.R
\name{write_data}
\alias{write_data}
\title{Function to extract analyser object data into a csv}
\usage{
write_data(object, which_data, type)
}
\arguments{
\item{object}{An object of class metime_plotter}

\item{which_data}{Character to specify the dataset}

\item{type}{which type of output file. Can be "csv", "tsv" and "xlsx"}
}
\value{
saves the data in the working directory as a csv and returns nothing
}
\description{
extracts information from analyser object and saves it as a csv
}
\examples{
see examples here
save_analyser_object(object, which_data="dataset")
}
\seealso{
\link{write_results}, \link{write_report}
}
