% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_report.R
\name{write_report}
\alias{write_report}
\title{Generates a report from an metime_analyser object}
\usage{
write_report(
  object,
  title = NULL,
  file = NULL,
  write_results = F,
  device = "html",
  interactive = F,
  author = NULL
)
}
\arguments{
\item{object}{a S4 object of class metime_analyser}

\item{title}{a character string used as title of the report. By default is set to metime + system time.}

\item{file}{a character string used as file name of the output report and table. By default is set to metime + system time.}

\item{write_results}{a logical indicating whether you want to save the plot data as an xlsx file. If TRUE an xlsx file with the file name specified by the title will be saved.}

\item{device}{a character string specifying the format of the report. By default is set to html. Other options include pdf.}

\item{interactive}{a logical indicating whether plots are interactive - only possible for html. Default set to FALSE}

\item{author}{a character specifying the name of the author}
}
\value{
Saves the report as html/pdf
}
\description{
Write an HTML or PDF report that summarizes one or more "me_time_analyser" objects and display all results
}
\seealso{
\link{write_report}, \link{write_results}
}
