% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_results.R
\name{write_results}
\alias{write_results}
\title{Function to extract results into different types of files}
\usage{
write_results(object, results_index, type)
}
\arguments{
\item{object}{An object of class metime_analyser}

\item{results_index}{character or numeric to define the results of interest}

\item{type}{character to define outfile type that is "csv", "xlsx" or "tsv"}
}
\value{
saves the data into a csv and returns nothing
}
\description{
extracts results from analyser object and saves it
}
\examples{
see examples here and maintain the order
If type is xlsx one out file is enough and the network data will be stored in different sheets
Network : save_results(object, results_index, out=c("edge", "node", "meta"), type="csv")
Others : save_results(object, results_index, out=c("outfile1", ...), type="tsv")
}
