% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data.R
\name{add_data}
\alias{add_data}
\title{Add a data frame to analyzer object.}
\usage{
add_data(object, which_data, type, x, id)
}
\arguments{
\item{object}{a S4 object of the class "metime_analyzer".}

\item{which_data}{a character to define which dataset is to be used.}

\item{type}{a character defining the data which should be appended. Can only be data, col_data, row_data. Default set to data.}

\item{x}{a data.frame that is merged to the data.}

\item{id}{a character vector defining the column to be used for matching. Default set to NULL (first column of x will be used as id column)}
}
\value{
a S4 object of class "metime_analyser" with appended data.frame x
}
\description{
Add a data.frame x to existing data within an analyzer object. Data can be added to data, col_data or row_data. A full join of the data is used.
}
