% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_distribution_vars_to_rows.R
\name{add_distribution_vars_to_rows}
\alias{add_distribution_vars_to_rows}
\title{Add phenotypic measurements that are not added to the row_data of the dataset}
\usage{
add_distribution_vars_to_rows(
  object,
  screening_vars,
  distribution_vars,
  which_data
)
}
\arguments{
\item{object}{An object of class metime_analyser}

\item{screening_vars}{A character vector to define the vars that are to be updated as per add_screening_vars() else set it to NULL(Default).}

\item{distribution_vars}{A character naming the vars of interest}

\item{which_data}{dataset to which the information is to be added(only 1 can be used at a time)}
}
\value{
object of class metime_analyser with phenotype data added to row data
}
\description{
A method applied on the s4 object of class "metime_analyser" to add all those datapoints that are present in phenotype dataframe but not in row_data
comes with the feature of updating those data points measured only at screening to all datapoints and then adding it to row_data
}
\examples{
object <- add_distribution_vars_to_rows(object=data, screening_vars=c("var1", "var2"), 
		distribution_vars=c("var1", "var2", "var3"), which_data="dataset1")
}
\seealso{
\link{add_screening_vars}
}
