% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_metabs_as_covariates.R
\name{add_metabs_as_covariates}
\alias{add_metabs_as_covariates}
\title{Function to add metabolites as covariates for network construction}
\usage{
add_metabs_as_covariates(object, which_data, which_metabs)
}
\arguments{
\item{object}{A S4 object of class metime_analyser}

\item{which_data}{Dataset to which the metab data is to be added(please note that this a single character)}

\item{which_metabs}{list of names of metabs and name of the list represents the dataset from which
the metabs are to be acquired. eg: which_metabs=list(nmr_data=c("metab1", "metab2"), lipid_data=c("metab3", "metab4"))}
}
\value{
S4 object with metabs from one dataset added for analysis to another dataset
}
\description{
Method applied on metime_analyser object to add other metabolite data to a certain dataset
}
