% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_node_features.R
\name{add_node_features}
\alias{add_node_features}
\title{Function to add features to visnetwork plot from another plotter object}
\usage{
add_node_features(object, results_indices, which_calculation = 1)
}
\arguments{
\item{object}{a S4 object of the class "metime_analyzer".}

\item{results_indices}{indices as a list to define which results to use. Eg: list(network=1/"name of the results", guide=2)}

\item{which_calculation}{index for plot_data to be used. Set to 1 by default.}
}
\value{
network plotter object with new node colors/features
}
\description{
Function to add node features to see the nodes in the network that affected differently
}
