% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_add.R
\name{add_screening_vars}
\alias{add_screening_vars}
\title{Function to add measurements taken at screening time for samples to be added to all timepoints}
\usage{
add_screening_vars(object, vars)
}
\arguments{
\item{object}{An object of class metab_analyser}

\item{vars}{A character naming the vars of interest}
}
\value{
phenotype data which can be replaced into the original object or use it separately with a different object
}
\description{
A method applied on the s4 object of class "metab_analyser" to add all those datapoints that were measured only during screening
to all the respective samples at all timepoints
}
