% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_general.R
\name{add_screening_vars}
\alias{add_screening_vars}
\title{Function to add measurements taken at screening time for samples to be added to all timepoints}
\usage{
add_screening_vars(object, vars)
}
\arguments{
\item{object}{An object of class metab_analyser}

\item{vars}{A character naming the vars of interest}
}
\value{
phenotype data which can be replaced into the original object or use it separately with a different object
}
\description{
Function to add measurements taken at screening time for samples to be added to all timepoints
}
