% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_calc.R
\name{calc_boruta}
\alias{calc_boruta}
\title{Function to calculate metabotype conservation index}
\usage{
calc_boruta(object, which_x, which_y, verbose, output_loc, file_name)
}
\arguments{
\item{object}{An object of class metab_analyser}

\item{verbose}{Information provided on steps being processed}

\item{which_data}{Name of the dataset to be used}
}
\value{
List of conservation index results
}
\description{
Function to calculate metabotype conservation index
}
