% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_calc.R
\name{calc_conservation_metabolite}
\alias{calc_conservation_metabolite}
\title{Function to calculate metabolite conservation index}
\usage{
calc_conservation_metabolite(object, which_data, timepoints, verbose)
}
\arguments{
\item{object}{An object of class metime_analyser}

\item{which_data}{Name of the dataset to be used}

\item{timepoints}{character vector with timepoints of interest}

\item{verbose}{Information provided on steps being processed}
}
\value{
List of conservation index results
}
\description{
Method applied on the object metime_analyser to calculate the metabotype conservation index
}
\examples{
#calculating metabolite_conservation_index 
out <- calc_metabolite_conservation(object=metime_analyser_object, which_data="Name of the dataset")
}
