% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_conservation_metabolite.R
\name{calc_conservation_metabolite}
\alias{calc_conservation_metabolite}
\title{Function to calculate metabolite conservation index}
\usage{
calc_conservation_metabolite(
  object,
  which_data,
  verbose,
  cols_for_meta,
  stratifications,
  name
)
}
\arguments{
\item{object}{An object of class metime_analyser}

\item{which_data}{Name of the dataset to be used}

\item{verbose}{Information provided on steps being processed}

\item{cols_for_meta}{A list of a Character vector to define column names that are to be used for plotting purposes}

\item{stratifications}{list to stratify the data used}

\item{name}{character vector to define the name of the results generated. length should be equal to which_data}
}
\value{
conservation index results that are added to the object
}
\description{
Method applied on the object metime_analyser to calculate the metabotype conservation index
}
\examples{
#calculating metabolite_conservation_index 
out <- calc_metabolite_conservation(object=metime_analyser_object, which_data="Name of the dataset")
}
