% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_calc.R
\name{calc_dimensionality_reduction}
\alias{calc_dimensionality_reduction}
\title{Function to calculate dimensionality reduction methods such as tsne, umap and pca.}
\usage{
calc_dimensionality_reduction(object, which_data, type)
}
\arguments{
\item{object}{An object of class metime_analyser}

\item{which_data}{a character vector - Names of the dataset from which the samples will be extracted}

\item{type}{type of the dimensionality reduction method to be applied. Accepted inputs are "UMAP", "tSNE", "PCA"}
}
\value{
a list with two dataframes containing the dimensionality reduction information 1) samples - data of the individuals(".$samples")
2) metabs - data of the metabolites(".$metabs")
}
\description{
A method to apply on s4 object of class metime_analyse in order to obtain information after dimensionality reduction on a dataset/s
}
\examples{
#calculate PCA
pca <- calc_dimensionality_reduction(object=metime_analyser_object, which_data="name/s of the dataset/s", type="PCA")
#calculate UMAP
pca <- calc_dimensionality_reduction(object=metime_analyser_object, which_data="name/s of the dataset/s", type="UMAP")
#calculate tSNE
pca <- calc_dimensionality_reduction(object=metime_analyser_object, which_data="name/s of the dataset/s", type="tSNE")
}
