% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gamm.R
\name{calc_gamm}
\alias{calc_gamm}
\title{Function to perform GAMM analysis}
\usage{
calc_gamm(
  object,
  which_data,
  stratifications = NULL,
  cols_for_meta = NULL,
  verbose = T,
  name = "regression_gamm",
  ...
)
}
\arguments{
\item{which_data}{Name of the dataset to be used}

\item{stratifications}{List to stratify data into a subset. Usage list(name=value)}

\item{cols_for_meta}{Character vector to define column names that are to be used for plotting purposes}

\item{verbose}{Information provided on steps being processed}

\item{name}{character vector to define the results. Should be equal to length of which_data}

\item{Object}{An S4 object of class metime_analyser}
}
\description{
An S4 method to perform GAMM analysis on a particular dataset. See regression_example to use a curated dataset.
}
