% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gamm.R
\name{calc_gamm}
\alias{calc_gamm}
\title{Function to perform Generalized additive models}
\usage{
calc_gamm(object, which_data, formula, cores)
}
\arguments{
\item{object}{An S4 object of class metime_analyser}

\item{which_data}{Dataset to be used for this analysis}

\item{formula}{A dataframe listing the formulae of the gamms to be used}

\item{cores}{number of cores of the system to be used. Can also be set to NULL}
}
\value{
plotter object with GAM results
}
\description{
Function to perform Generalized additive models
}
