% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ggm_genenet.R
\name{calc_ggm_genenet}
\alias{calc_ggm_genenet}
\title{An automated fucntion to calculate GGM from genenet crosssectional version}
\usage{
calc_ggm_genenet(
  object,
  which_data,
  threshold,
  all,
  cols_for_meta,
  covariates,
  stratifications,
  name,
  ...
)
}
\arguments{
\item{object}{S4 object of cĺass metime_analyser}

\item{which_data}{a character or a character vector naming the datasets of interest}

\item{threshold}{type of threshold to be used for extracting significant edges.
allowed inputs are "li", "FDR", "bonferroni"}

\item{all}{Logical to extract all edges without any pval correction}

\item{cols_for_meta}{list of character vector for extracting metadata of metabolites for plotting}

\item{covariates}{covariates to be used for this analysis}

\item{stratifications}{List to stratify data into a subset. Usage list(name=value)}

\item{name}{Name of the result}

\item{...}{additional arguments for GeneNet}
}
\value{
Network data as a plotter object
}
\description{
automated funtion that can be applied on metime_analyser object to obtain geneNet network along with threshold used
}
