% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ggm_multibipartite_lasso.R
\name{calc_ggm_multibipartite_lasso}
\alias{calc_ggm_multibipartite_lasso}
\title{An automated fucntion to calculate GGM from multibipartite lasso approach}
\usage{
calc_ggm_multibipartite_lasso(
  object,
  which_data,
  alpha = 1,
  nfolds = 3,
  stratifications,
  cols_for_meta,
  cores = 4
)
}
\arguments{
\item{object}{S4 object of cĺass metime_analyser}

\item{which_data}{a character or a character vector naming the datasets of interest}

\item{alpha}{tuning parameter for lasso + ridge regression in glmnet. Default set to 1 to perform LASSO}

\item{nfolds}{nfolds for cv.glmnet. Default set to 3}

\item{stratifications}{List to stratify data into a subset. Usage list(name=value)}

\item{cols_for_meta}{a list of character vectors of column names to be used for visualization of the networks.}

\item{cores}{number of cores to be used in parallel::mclapply(). Default set to 4.}

\item{name}{character to define the name of the results}

\item{...}{additional arguments for cv.glmnet function}
}
\value{
Analyser object with updated results of this calculation
}
\description{
automated funtion that can be applied on s4 object of class metime_analyser to calculate a network using
multibipartite lasso
}
