% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ggm_multibipartite_lasso.R
\name{calc_ggm_multibipartite_lasso}
\alias{calc_ggm_multibipartite_lasso}
\title{An automated fucntion to calculate GGM from multibipartite lasso approach}
\usage{
calc_ggm_multibipartite_lasso(
  object,
  which_data,
  alpha,
  nfolds,
  timepoints,
  cols_for_meta
)
}
\arguments{
\item{object}{S4 object of cĺass metime_analyser}

\item{which_data}{a character or a character vector naming the datasets of interest}

\item{alpha}{tuning parameter for lasso + ridge regression in glmnet}

\item{nfolds}{nfolds for cv.glmnet}

\item{timepoints}{timepoints of interest that are to be used to build networks(as per timepoints in rows)}

\item{cols_for_meta}{a list of character vectors of column names to be used for visualization of the networks.}
}
\value{
list of plotter objects that can be used for plotting.
}
\description{
automated funtion that can be applied on s4 object of class metime_analyser to calculate a network using
multibipartite lasso
}
