% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_calc.R
\name{calc_ggm_multibipartite_lasso}
\alias{calc_ggm_multibipartite_lasso}
\title{An automated fucntion to calculate GGM from multibipartite lasso approach}
\usage{
calc_ggm_multibipartite_lasso(object, which_data, alpha, nfolds, timepoints)
}
\arguments{
\item{object}{S4 object of cĺass metime_analyser}

\item{which_data}{a character or a character vector naming the datasets of interest}

\item{alpha}{tuning parameter for lasso + ridge regression in glmnet}

\item{nfolds}{nfolds for cv.glmnet}

\item{timepoints}{timepoints of interest that are to be used to build networks(as per timepoints in rows)}
}
\value{
Network data with edges and their respective betas
}
\description{
automated funtion that can be applied on s4 object of class metime_analyser to calculate a network using
multibipartite lasso
}
