% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_lm.R
\name{calc_lm}
\alias{calc_lm}
\title{Cossectional linear models (per time point)}
\usage{
calc_lm(
  object,
  which_data,
  stratifications = NULL,
  cols_for_meta = NULL,
  threshold = c("none", "nominal", "li", "fdr", "bonferroni"),
  verbose = T,
  name = "regression_lm_1",
  timepoint = NULL,
  num_cores = NULL
)
}
\arguments{
\item{object}{an S4 object of class metime_analyser}

\item{which_data}{a character defining the name of the dataset to be used.}

\item{stratifications}{list to stratify data into a subset. Usage list(name=value). Default set to NULL, thereby not performing any type of stratification.}

\item{cols_for_meta}{a character vector to define column names that are to be used for plotting purposes. Default set to NULL, therby not adding columns as metadata.
If you want automated facet wrapping option then set your new_columns as "facet_your_name"}

\item{threshold}{a character vector to define the type of threshold for significant interactions. Default set to all availabe thresholds: c("none","nominal","li","fdr","bonferroni").
allowed inputs are "li", "FDR", "bonferroni" and "nominal"(cutoff p=0.05, set as Default)}

\item{verbose}{a logical on whether to print the calculation progress. Default set to FALSE.}

\item{name}{a character vector to define the index within the results. Should be equal to length of which_data. Default set to regression_lm_1.}

\item{timepoint}{time input for cross-sectional model should be the same as the value in time column in row_data.}

\item{num_cores}{numeric input to define the number of cores that you want to use for parallel computing. Default is set to NULL which is parallel::detectCores() -1.}
}
\value{
a S4 object of the class metime_analyzer with analysis results appended to the result section.
}
\description{
Fit multiple linear models (lm) onto one dataset. Covariates can be added using the covariate column of the col_data (multiple covariates can be added, separated by '###') See examples for more information.
}
\details{
Add details here
}
