% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_lm.R
\name{calc_lm}
\alias{calc_lm}
\title{Calculation of crossectional linear models (per time point)}
\usage{
calc_lm(
  object,
  which_data,
  stratifications = NULL,
  cols_for_meta = NULL,
  threshold = "nominal",
  verbose = T,
  name = "regression_lm_1"
)
}
\arguments{
\item{object}{an S4 object of class metime_analyser}

\item{which_data}{a character defining the name of the dataset to be used.}

\item{stratifications}{list to stratify data into a subset. Usage list(name=value). Default set to NULL, thereby not performing any type of stratification.}

\item{cols_for_meta}{a character vector to define column names that are to be used for plotting purposes. Default set to NULL, therby not adding columns as metadata.}

\item{threshold}{a character of length 1 to define the type of threshold for significant interactions.
allowed inputs are "li", "FDR", "bonferroni" and "nominal"(cutoff p=0.05, set as Default)}

\item{verbose}{a logical on whether to print the calculation progress. Default set to FALSE.}

\item{name}{a character vector to define the index within the results. Should be equal to length of which_data. Default set to regression_lm_1.}
}
\value{
a S4 object of the class metime_analyzer with analysis results appended to the result section.
}
\description{
Fits multiple linear models (LM).
}
\details{
Add details here
}
