% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_calc.R
\name{calc_lmm}
\alias{calc_lmm}
\title{Function to perform Linear Mixed Models}
\usage{
calc_lmm(object, which_data, formula = NULL, cores = NULL)
}
\arguments{
\item{object}{An S4 object of class metime_analyser}

\item{which_data}{Dataset to be used}

\item{formula}{Formulae to define the equation to be used in linear mixed models}

\item{cores}{Numeric to define the number of cores to be used}
}
\value{
Returns a plotter object for plotting forest plots of the results
}
\description{
Function to perform linear mixed models on dataset of interest
}
