% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_lmm.R
\name{calc_lmm}
\alias{calc_lmm}
\title{Calculation of linear mixed models}
\usage{
calc_lmm(
  object,
  which_data,
  stratifications = NULL,
  cols_for_meta = NULL,
  random = "subject",
  threshold = c("none", "nominal", "li", "fdr", "bonferroni"),
  interaction = NULL,
  verbose = T,
  name = "regression_lmm_1"
)
}
\arguments{
\item{object}{an S4 object of class metime_analyser}

\item{which_data}{a character defining the name of the dataset to be used.}

\item{stratifications}{list to stratify data into a subset. Usage list(name=value). Default set to NULL, thereby not performing any type of stratification.}

\item{cols_for_meta}{a list of character vector to define column names that are to be used for plotting purposes. Default set to NULL, therby not adding columns as metadata.
If you want automated facet wrapping option then set your new_columns as "facet_your_name"}

\item{random}{a character vector defining which variables should be treated as random effects. Default set to "subject".}

\item{threshold}{a character of length 1 to define the type of threshold for significant interactions.
allowed inputs are "li", "FDR", "bonferroni" and "nominal"(cutoff p=0.05, set as Default)}

\item{interaction}{a character vector defining which interaction terms should be added to the model. Default set to NULL, with no interaction added.}

\item{verbose}{a logical on whether to print the calculation progress. Default set to FALSE.}

\item{name}{a character vector to define the index within the results. Should be equal to length of which_data. Default set to regression_gamm_1.}
}
\value{
a S4 object of the class metime_analyzer with analysis results appended to the result section.
}
\description{
Fits multiple Linear mixed models (LMM) to a longitudinal dataset.
}
\details{
Add details here
}
