% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_parafac.R
\name{calc_parafac}
\alias{calc_parafac}
\title{Function to perform PARAFAC analysis}
\usage{
calc_parafac(object, which_data, timepoints, nfac = 3, ...)
}
\arguments{
\item{object}{S4 object of class metime_analyser}

\item{which_data}{character vector for dataset to be used}

\item{nfac}{parameter nfac for parafac(). Numeric value to define the number of factors. Default is set to 3}

\item{...}{Additional arguments to be used for the function parafac()}

\item{stratifications}{list to define stratification for the data before calculation}
}
\value{
An object of class PARAFAC. See multiway library for more information
}
\description{
Function for S4 object of class metime_analyser to perform PARAFAC analysis
}
