% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_temporal_network.R
\name{calc_temporal_network}
\alias{calc_temporal_network}
\title{An automated function to caluclate temporal network with lagged model}
\usage{
calc_temporal_network(
  object,
  which_data,
  lag,
  stratifications,
  alpha = 1,
  nfolds = 3,
  cols_for_meta,
  names
)
}
\arguments{
\item{object}{S4 object of class metime_analyser}

\item{which_data}{dataset or datasets to be used}

\item{lag}{which lagged model to use. 1 means one-lagged model, similarly 2,3,..etc}

\item{stratifications}{List to stratify data into a subset. Usage list(name=value)}

\item{alpha}{parameter for regression coefficient. Set to 1 for lasso regression}

\item{nfolds}{nfolds parameter for glmnet style of regression. Default is set to 3}

\item{cols_for_meta}{a list of character vectors of column names to be used for visualization of the networks.}

\item{names}{character vector with the same length as that of possible models}
}
\value{
S4 object with updated temporal network results
}
\description{
calculates temporal networks for each dataset with a lagged model as used in graphical VAR
}
