% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_calc.R
\name{calc_wgcna_eigenmetabolites}
\alias{calc_wgcna_eigenmetabolites}
\title{Function to extract eigenmetabolites and modules from WGCNA}
\usage{
calc_wgcna_eigenmetabolites(object, which_data, baseline, ...)
}
\arguments{
\item{object}{An s4 object of class metime_analyser}

\item{which_data}{Dataset to be used for the analysis}

\item{baseline}{baseline timepoint}
}
\value{
plotter object with dendogram results
}
\description{
Function to calculate modules and eigenmetabolites
}
