% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get.R
\name{get_class_info_from_edges}
\alias{get_class_info_from_edges}
\title{Function to get information on how many class edges are present}
\usage{
get_class_info_from_edges(calc_networks, metadata)
}
\arguments{
\item{calc_networks}{list of calculated networks}

\item{metadata}{metadata of the edges present}
}
\value{
table with information on different type of edges present
}
\description{
Function to check how the different edges in a GGM are associated to their
respective classes(it could be super-pathway or sub-pathway)
}
