% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get.R
\name{get_coldata}
\alias{get_coldata}
\title{Function to extract col data of a dataset}
\usage{
get_coldata(object, which_data)
}
\arguments{
\item{object}{An object of class S4}

\item{which_data}{Dataset of interest}
}
\value{
col data of the dataset of interest
}
\description{
Function to get coldata
}
