% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_common_samples_at_timepoints.R
\name{get_common_samples_at_timepoints}
\alias{get_common_samples_at_timepoints}
\title{Function to get common samples at multiple timepoints chosen}
\usage{
get_common_samples_at_timepoints(object, which_data, timepoints)
}
\arguments{
\item{object}{An S4 object of class metime_analyser}

\item{which_data}{Dataset/s of interest}

\item{timepoints}{A character vector to define timepoints}
}
\value{
character vector of common subjects at common timepoints
}
\description{
Get a vector of common samples that can be used for stratification
}
