% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Function to get particular dataset's data}
\usage{
get_data(object, which_data)
}
\arguments{
\item{object}{An S4 object of class metime_analyser}

\item{which_data}{character/index to define the data of interest}
}
\value{
data.frame with data of interest
}
\description{
function to get data from S4 object
}
