% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get.R
\name{get_files_and_names}
\alias{get_files_and_names}
\title{Function to pack all the data into a single object of class "metab_analyser"}
\usage{
get_files_and_names(path, annotations_index)
}
\arguments{
\item{path}{Path to the parent directory}

\item{annotations_index}{a list to be filled as follows = list(phenotype="Name or index of the files", medication="Name or index of the files")}
}
\value{
An object of class metab_analyser
}
\description{
This function loads all the files from the parent directory. It assumes a
certain naming pattern as follows: "datatype_\link{NULL|col|row}_data.rds"
Any other naming pattern is not allowed. The function first writes
all files into a list and each type of data is packed into its respective
class i.e. col_data, row_data or data
}
