% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_make_results.R
\name{get_make_results}
\alias{get_make_results}
\title{Function to make results list for metime_analyser object}
\usage{
get_make_results(object, data, metadata, calc_type, calc_info, name)
}
\arguments{
\item{object}{An S4 object of class metime_analyser}

\item{data}{list of dataframes of plotable data obtained from any calc function}

\item{metadata}{dataframe or a list of dataframes with the metadata for the plot table mentioned above. To obtain these see
get_metadata_for_rows() and get_metadata_for_columns()}

\item{calc_type}{A character vector to specify type of calculation - will be used for comp_ functions
For networks the accepted notations are "genenet_ggm", "multibipartite_ggm", and "temporal_network"
sjould be the same length as the list of data provided}

\item{calc_info}{A string to define the information about calculation, should be the same length as the list
data provided}

\item{name}{Name of the result}
}
\value{
object with results of the calculation updated
}
\description{
function to generate results for metime_analyser object
}
