% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get.R
\name{get_metadata_for_columns}
\alias{get_metadata_for_columns}
\title{Get metadata for columns(in most cases for metabolites)}
\usage{
get_metadata_for_columns(object, which_data, columns, names, index_of_names)
}
\arguments{
\item{object}{S4 object of class MeTime Analyser}

\item{which_data}{Names of dataset/s to be used}

\item{columns}{A list of character vectors for the columns of interest. Length of the list should be
same as length of which_data}

\item{names}{A Character vector with the new names for the columns mentioned above id should always be first in order}

\item{index_of_names}{character vector to define the name of the column in which names of the variables are stored}
}
\value{
data.frame with metadata information
}
\description{
function to generate a metadata list for building the MeTime plotter object
}
