% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get.R
\name{get_metadata_for_plotting}
\alias{get_metadata_for_plotting}
\title{Function to extract metadata of the metabolites and samples for visualization(both dimensionality reduction and GGMs)}
\usage{
get_metadata_for_plotting(
  object,
  which_data,
  metab_groups,
  metab_ids,
  cols_for_vis_samples,
  screening_vars
)
}
\arguments{
\item{object}{S4 object of class metab_analyse}

\item{which_data}{choose the dataset from which metabolites will be extracted for metadata}

\item{metab_groups}{choose the column that has metabolite groups}

\item{metab_ids}{chodse the column that has metabolite names}

\item{cols_for_vis_samples}{character vector representing the name of the columns in phenotype data}

\item{screening_vars}{character vector representing the measurements obtained at baseline to be added to all timepoints for visualization.
is set to NULL if nothing is added to it}
}
\value{
metadata dataframe with names, groups and class
}
\description{
A method applied onto s4 object of class "metab_analyser" so as to obtain metadata of the metabolites and samples.
Metadata includes their ontology that is the pathway they belong to and also the class or the dataset type. can also add colors
for the metabolites for visualization as a separate column. For samples the metadata is basically the columns of interest from the phenotype table
that can be used to see sample information in the interactive plot.
}
\examples{
# metadata_list <- get_metadata_for_plotting(object=metab_analyser_object, which_data="name/s of datasets", metab_groups="colname/s of the group column in each dataset in order"
#										metab_ids="colname/s of the metabolite_name column in each dataset in order", cols_for_vis_samples="colnames of phenotype data for samples", screeing_vars=TRUE/FALSE)
}
