% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get.R
\name{get_metadata_for_rows}
\alias{get_metadata_for_rows}
\title{Get metadata for rows(in most cases for samples)}
\usage{
get_metadata_for_rows(object, which_data, columns, names)
}
\arguments{
\item{object}{S4 object of class MeTime Analyser}

\item{which_data}{Names of dataset/s to be used}

\item{columns}{A list of character vectors for the columns of interest. Length of the list should be
same as length of which_data}

\item{names}{A Character vector with the new names for the columns mentioned above}
}
\value{
data.frame with metadata information for rows
}
\description{
function to generate a metadata list for building the MeTime plotter object
}
