% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_general.R
\name{get_palette}
\alias{get_palette}
\title{Function to get a palette of distinct colorblind friendly colors.}
\usage{
get_palette(n)
}
\arguments{
\item{n}{number of colors wanted in the palette}
}
\value{
a color palette vector with colors in the form of hex codes
}
\description{
Function to get a palette of distinct colorblind friendly colors.
}
