% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters_of_results.R
\name{get_parameters_of_results}
\alias{get_parameters_of_results}
\title{Get parameters of the results}
\usage{
get_parameters_of_results(object, results_index)
}
\arguments{
\item{object}{An S4 object of class metime_analyser.}

\item{results_index}{a character or numeric value to be used as the result index.}
}
\value{
a list with all the function names and the values of the arguments passed into it
}
\description{
Get parameters and functions applied to obtain results.
}
