% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters_of_results.R
\name{get_parameters_of_results}
\alias{get_parameters_of_results}
\title{Get parameters of the results}
\usage{
get_parameters_of_results(object, results_index)
}
\arguments{
\item{object}{An S4 object of class metime_analyser}

\item{results_index}{name or index to get to the results of interest}
}
\value{
a dataframes with functions(rownames) and parameters(colnames)
}
\description{
Function to get parameters and functions applied to obtain results
}
