% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{get_results}
\alias{get_results}
\title{Function to get results of a particular analysis}
\usage{
get_results(object, which_result)
}
\arguments{
\item{object}{An S4 object of class metime_analyser}

\item{which_result}{character/index to define the result of interest}
}
\value{
list with results data list and information regarding
the functions applied and type of calculation
}
\description{
function to get results from S4 object
}
