% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rowdata.R
\name{get_rowdata}
\alias{get_rowdata}
\title{Get row_data from a S4 object of the class "metime_analyser"}
\usage{
get_rowdata(object, which_data)
}
\arguments{
\item{object}{a S4 object of the class "metime_analyser"}

\item{which_data}{a character of the dataset name of interest.}
}
\value{
A dataframe containing the row data of the dataset of interest.
}
\description{
Get the row information of a specified dataset from a S4 object of the class "metime_analyser"
}
