% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_samples_and_timepoints.R
\name{get_samples_and_timepoints}
\alias{get_samples_and_timepoints}
\title{Summarize number of time points and the total number of samples available at that point}
\usage{
get_samples_and_timepoints(object, which_data)
}
\arguments{
\item{which_data}{a character to define which dataset is to be used.}

\item{objecta}{a S4 object of the class "metime_analyzer".}
}
\value{
A dataframe with time points and number of samples at each time point.
}
\description{
Obtain the number of unique samples available at each time point.
}
