% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_get.R
\name{get_samples_and_timepoints}
\alias{get_samples_and_timepoints}
\title{Function to know the number of timepoints and the total number of samples available at that point}
\usage{
get_samples_and_timepoints(object, which_data)
}
\arguments{
\item{object}{An object of class metab_analyser}

\item{which_data}{Name of the dataset in context}
}
\value{
A data table with timepoints and number of samples at each timepoint
}
\description{
A method applied onto s4 object of class "metab_analyser" so as to obtain the number of unique samples available
at each timepoint.
}
