% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_importer.R
\name{get_stratified_data}
\alias{get_stratified_data}
\title{Function to stratify before calculation}
\usage{
get_stratified_data(object, which_data, stratifications)
}
\arguments{
\item{object}{An S4 object of class metime_analyser}

\item{which_data}{character vector to define Dataset/s to be used}

\item{stratifications}{list of variables and their values to stratified - list(name=value).
example of stratification with time stratifications=list(time=c("timepoint1", "timepoint2", ...)) then only these timepoints
will be considered in the analysis}
}
\value{
data_list that contains col_data, row_data and data with stratifications aforementioned. Access them as shown:
data=data_list[\link{"data"}]; col_data=data_list[\link{"col_data"}]; row_data=data_list[\link{"row_data"}]
}
\description{
get function to perform stratification of a dataset in calculations. Used only in calc_ functions.
}
