% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_regression.R
\name{meta_regression}
\alias{meta_regression}
\title{Comparison of regression analysis}
\usage{
meta_regression(
  object,
  method = c("sign", "cor", "het"),
  result_index = NULL,
  name = "meta_regression_1",
  update_table = T
)
}
\arguments{
\item{object}{a S4 object of class metime_analyser.}

\item{method}{a character vector of methods 'sign' (sign of beta), 'het' (heterogeneity), 'cor' (correlation). See details for more information.}

\item{result_index}{a character vector of results to compare. Default set to NULL. If set to NULL all results from regression are used.}

\item{name}{a character containing the name of the new result element appended to the analyzer object. Default set to 'comp_regression_1'.}

\item{update_table}{a logical whether the results should be added to the plotting section. If true tables will be printed in the report.}
}
\value{
a S4 object of class metime_analyser object with function output appended to results
}
\description{
A function to compare results from regression analysis (LM, LMM, GAMM)
}
\details{
Multiple options are available to compare betas and betas + se from regression analysis. \n sign: comparison of beta signs. \n het (heterogeneity): Significance of heterogeneity across two models as described by \href{\url{https://doi.org/10.1111/j.1745-9125.1998.tb01268.x}}{Paternoster, 1998} and \href{\url{https://doi.org/10.1371/journal.pgen.1002215}}{Mittelstrass, 2011} . \n cor (correlation): Similarity of betas based on Pearson correlation
}
\seealso{
\link{meta_parametric_test}, \link{meta_nonparametric_test}, \link{meta_network}
}
