% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeTime-package.R
\docType{package}
\name{MeTime-package}
\alias{MeTime}
\alias{MeTime-package}
\title{MeTime: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Bharadwaj Marella \email{bharadwaj.marella@helmholtz-muenchen.de}

Authors:
\itemize{
  \item Patrick Dreher \email{patrick.dreher@helmholtz-muenchen.de}
}

}
\keyword{internal}
