% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_extract_common_samples.R
\name{mod_extract_common_samples}
\alias{mod_extract_common_samples}
\title{Function to get only common samples from the dataframes in list_of_data}
\usage{
mod_extract_common_samples(object)
}
\arguments{
\item{object}{An object of class metime_anaylser}
}
\value{
list_of_data with common samples across all time points
}
\description{
A method applied on object of class metime_analyse to extract common samples across datasets.
}
\examples{
# extracting common samples across all datasets
new_list_of_data <- mod_extract_common_samples(object=metime_analyser_object)
}
