% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filter.R
\name{mod_filter}
\alias{mod_filter}
\title{Modification function to filter columns in data, row_data or col_data}
\usage{
mod_filter(object, which_data, type = "data", ...)
}
\arguments{
\item{object}{A S4 object of class metime_analyser.}

\item{which_data}{a vector of character defining which data should be merged.}

\item{type}{a character defining which data to filter. Can be "data","row_data", or "col_data". Default set to "data". Note: if a dataset is filtered the row_data will be changed accordingly}

\item{...}{arguments to pass directly into dplyr::filter() function.}
}
\value{
object with mutated data, col_data or row_data
}
\description{
Modification (mod) function to filter columns in data, row_data or col_data
}
