% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filter_timepoints.R
\name{mod_filter_timepoints}
\alias{mod_filter_timepoints}
\title{Functions for selecting time points}
\usage{
mod_filter_timepoints(object, timepoints, complete, which_data)
}
\arguments{
\item{object}{An object of class metime_analyser}

\item{timepoints}{time points to be selected.}

\item{complete}{if TRUE subjects are only selected if measured in all selected time points}

\item{which_data}{Name of the dataset to be used}
}
\value{
An object of class metime_analyser with processed data
}
\description{
a method applied onto class metime_analyser in order to extract timepoints of interest from a dataset
}
\examples{
#example to use this function
object <- mod_filter_tp(object, timepoints=c("t0","t12","t24"), full=TRUE, which_data="Name of the dataset")
}
