% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_mod.R
\name{mod_filter_tp}
\alias{mod_filter_tp}
\title{Functions for selecting time points}
\usage{
mod_filter_tp(object, timepoints, full, which_data)
}
\arguments{
\item{object}{An object of class metime_analyser}

\item{timepoints}{time points to be selected}

\item{full}{if TRUE subjects are only selected if measured in all selected time points}

\item{which_data}{Name of the dataset to be used}
}
\value{
An object of class metime_analyser with processed data
}
\description{
a method applied onto class metime_analyser in order to extract timepoints of interest from a dataset
}
\examples{
#example to use this function
object <- mod_filter_tp(object, timepoints=c(0,12,24), full=TRUE, which_data="Name of the dataset")
}
