% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_importer.R
\name{mod_generate_plots}
\alias{mod_generate_plots}
\title{Function to update plots post calculations}
\usage{
mod_generate_plots(object, .interactive = FALSE, type, results_index = NULL)
}
\arguments{
\item{object}{An S4 object of class metime_analyser}

\item{.interactive}{logical to make the plot interactive or not}

\item{type}{character to define the type of calculation used for updating the plot
cols_for_samples, cols_for_metabs, cols_for_meta etc will be used. So make sure you set those correctly
for better results.
Allowed inputs are: c("ggm|network", "PCA", "UMAP", "tSNE", "CI_metabotype", "CI_metabolite",
"pairwise_distance", "pairwise_correlation", "colinearity", "distribution_metabs", "distribution_samples", "regression", "feature_selection")}

\item{results_index}{character/numeric input to define the results that you want to plot or replot with our automation.
Default will be set to NULL. Length of results_index should be equal to 1.}
}
\value{
object with plots of the newest calculation
}
\description{
Modification(mod) Function to generate all possible/available plots of a calculation.
This function is a wrapper function for plot()
}
\seealso{
\link{plot}
}
