% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_impute.R
\name{mod_impute}
\alias{mod_impute}
\title{Imputation of missing values}
\usage{
mod_impute(
  object,
  which_data,
  method = "rf",
  thresh = 0.3,
  path_to_diagnostics = NULL
)
}
\arguments{
\item{object}{a S4 object of class "metime_analyser".}

\item{which_data}{a character defining which dataset should be imputed.}

\item{method}{a character defining the method for imputation. See details for more information. Default set to 'rf'}

\item{thresh}{a numeric defining at which missingness level metabolites will be excluded. Default set to 0.3 (30\% missingness)}

\item{path_to_diagnostics}{a character defining the path to where a diagnostic file is being written.}
}
\value{
S4 object of the class "metime_analyser" with mutated col_data and row_data.
}
\description{
This function imputes missing values using one of the implemented imputation methods.
}
\details{
additional methods are currently not implemented such as knn, pmm, norm etc.
}
