% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_mod.R
\name{mod_merge_data}
\alias{mod_merge_data}
\title{Function to merge two sets of data for any analysis}
\usage{
mod_merge_data(object, which_data, name)
}
\arguments{
\item{object}{An S4 object of class metime_analyser}

\item{which_data}{Datasets to be merged. Only two or more are allowed}

\item{name}{character vector to define the name of the new dataset}
}
\value{
A new S4 object of class metime_analyser with the new merged dataset appended to it
}
\description{
A function to merge two or more datasets and use it for analysis
}
