% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_merge_data.R
\name{mod_merge_data}
\alias{mod_merge_data}
\title{Modification (mod) function to merge two sets of data for any analysis}
\usage{
mod_merge_data(object, which_data, filter_samples, name)
}
\arguments{
\item{object}{a S4 object of class metime_analyser.}

\item{which_data}{a vector of character defining which data should be merged. Has to contain two or more values.}

\item{filter_samples}{a character specifying if samples should be filtered. Default set to no filtering. Other options include common samples (samples used if found in all datasets), or set to name of dataset (samples filtered by samples of one dataset)}

\item{name}{a character vector to define the name of a new dataset.}
}
\value{
a new S4 object of class metime_analyser with the  new merged dataset appended to it
}
\description{
a modification function that merges two sets of data including data, col_data and row_data.
}
