% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_merge_results.R
\name{mod_merge_results}
\alias{mod_merge_results}
\title{Function to combine plotter objects}
\usage{
mod_merge_results(object, results_indices, groups)
}
\arguments{
\item{object}{An S4 object of class metime_analyser}

\item{results_indices}{one or more indices of results to merge data}

\item{groups}{character vector to define the groups that are involved}
}
\value{
object with add merged letters
}
\description{
Function to combine plotter objects based on similar calc type
}
