% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mutate.R
\name{mod_mutate}
\alias{mod_mutate}
\title{Function to mutate columns in row_data or col_data}
\usage{
mod_mutate(object, which_data, type = "data", ...)
}
\arguments{
\item{object}{An S4 object of class metime_analyser}

\item{which_data}{Dataset/result of interest. Has to be of length=1}

\item{type}{either "row" for row_data or "col" for col_data or "data" for data and "results" for results.
Set to "data" as default}

\item{...}{arguments to pass directly into dplyr::mutate() function.}
}
\value{
object with mutated col_data and row_data
}
\description{
This function allows you to mutate columns by changing class or ...
}
