% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_split_acc_to_time.R
\name{mod_split_acc_to_time}
\alias{mod_split_acc_to_time}
\title{Function to split data acoording to time}
\usage{
mod_split_acc_to_time(object)
}
\arguments{
\item{object}{An object of class metime_analyser}
}
\value{
list_of_data with each dataframe being broken into a list of dataframes with respect to the timepoint they belong to
}
\description{
Function to split the list of dataframes into a nested list with each dataframe
being split into into dataframes of different timepoints
}
\examples{
#splitting data according to time
new_data <- mod_split_acc_to_time(object=metime_analyser_object)
}
