% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_trans_eigendata.R
\name{mod_trans_eigendata}
\alias{mod_trans_eigendata}
\title{Function to add the clusters obtained from wgcna}
\usage{
mod_trans_eigendata(
  object,
  which_data,
  append,
  results_index = NULL,
  cols_for_meta = NULL,
  name = "WGCNA_clusters_1",
  ...
)
}
\arguments{
\item{object}{An S4 object of class metime_analyser}

\item{which_data}{character to define which dataset is to be used}

\item{append}{logical if set to true adds the new data to the object used else creates new object}

\item{results_index}{results_index if clusters were previously calculated else set to NULL(default)}

\item{cols_for_meta}{A list of named character vector to extract col_data to add it for the eigendata. will be parsed to
get_metadata_for_columns}

\item{...}{arguments for calc_clusters_wgcna. Make sure to set the correct baseline value if you are using the function directly}
}
\value{
metime_analyser object with new dataset with eigendata of the metabolites
}
\description{
Modification(mod) function to get cluster assiginment and generate eigenmetabolite matrix from a dataset.
}
\seealso{
\link{get_metadata_for_columns}, \link{calc_clusters_wgcna}
}
