% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_trans_eigendata.R
\name{mod_trans_eigendata}
\alias{mod_trans_eigendata}
\title{Function to add the clusters obtained from wgcna}
\usage{
mod_trans_eigendata(object, which_data, append, clusters, ...)
}
\arguments{
\item{object}{An S4 object of class metime_analyser}

\item{which_data}{character to define which dataset is to be used}

\item{append}{logical if set to true adds the new data to the object used else creates new object}

\item{clusters}{logical if set to true will add already existing cluster info otherwise creates new}

\item{...}{arguments for add_clusters_wgcna}
}
\value{
metime_analyser object with new dataset with eigendata of the metabolites
}
\description{
A function to add cluster assiginment to the col_data from WGCNA
}
