% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_mod.R
\name{mod_trans_zscore}
\alias{mod_trans_zscore}
\title{Function to scale the data}
\usage{
mod_trans_zscore(object, which_data)
}
\arguments{
\item{object}{An object of class metime_analyser}

\item{which_data}{Name of the dataset to be used}
}
\value{
An object of class metime_analyser with processed data
}
\description{
Functions for scaling
}
\examples{
# example to apply scaling
object <- mod_zscore(object, which_data="name of the dataset")
}
