% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_distribution_plotter.R
\name{viz_distribution_plotter}
\alias{viz_distribution_plotter}
\title{Function for Plotting distributions of phenotypic variables}
\usage{
viz_distribution_plotter(object, colname, which_data, strats, phenotype)
}
\arguments{
\item{object}{An object of class metime_analyser}

\item{colname}{Name of the variable whose distribution is of interest}

\item{which_data}{Name of the dataset from which the samples will be extracted}

\item{strats}{Character vector with colnames that are to be used for stratification}

\item{phenotype}{Logical. If true data will be collected from phenotype_data matrix else from row data}
}
\value{
a list with either 1) density plot, mean table acc to timepoint and variable type or
2) bar plot, line plot, and variable type
}
\description{
A method to be applied onto s4 object so as to obtain distributions of various phenotypic variables
}
\examples{
# extracting distribuiton of Age from dataset1
plot <- viz_distribution_plotter(object, colname="Age", which_data="dataset1", strats="additional columns for facet wrapping")
}
