% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_viz.R
\name{viz_plotter_ggplot}
\alias{viz_plotter_ggplot}
\title{Setting up standard wrapper for all ggplot plots for a metime_plotter object.}
\usage{
viz_plotter_ggplot(object, aesthetics)
}
\arguments{
\item{object}{S4 object of class metime_plotter}

\item{aesthetics}{list for aesthetics. eg: list(list(x="colname",y="colname",color="colname", shape="colname"), list(...)) for "dot" plot and "heatmap"
plot, for heatmap: list(x="colname", y="colname", fill="colname"). Additionally two other character vectors are allowed namely .$vis and .$strats for text
and for facet wrapping.}
}
\value{
metime_plotter object with updated plot
}
\description{
plot function for metime_plotter object with different inputs to specialize plots. Used for all calc outputs.
}
