% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_results.R
\name{write_results}
\alias{write_results}
\title{Function to save plot_data from an analyser_objec into an xlsx file}
\usage{
write_results(object, file = NULL)
}
\arguments{
\item{object}{An object of class metime_analyser}

\item{file}{a character defining the path where to save the xlsx file}
}
\value{
saves all plot data in an xlsx sheet
}
\description{
Saves plot_data from an analyser_objec into an xlsx file (no Java)
}
\seealso{
\link{write_report}, \link{write_data}
}
